% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecheck.R
\name{add_class_hierarchy_cache}
\alias{add_class_hierarchy_cache}
\title{Add a Class Hierarchy to the Cache}
\usage{
add_class_hierarchy_cache(hierarchy)
}
\arguments{
\item{hierarchy}{\code{character(1)} the class hieararchy to add}
}
\description{
Add a class hierarchy to the class hierarchy cache.
}
\seealso{
Other class hierarchy operations: \code{\link{reset_class_hierarchy_cache}}
}
\concept{class hierarchy operations}
