% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_graph.R
\name{as_graph}
\alias{as_graph}
\title{Conversion to mlr3pipeline Graph}
\usage{
as_graph(x, clone = FALSE)
}
\arguments{
\item{x}{(\code{any}) \cr
Object to convert.}

\item{clone}{(\code{logical(1)}) \cr
Whether to return a (deep copied) clone if \code{x} is a Graph.}
}
\value{
\code{\link{Graph}} \code{x} or a deep clone of it.
}
\description{
The object is turned into a \code{Graph} if possible.
If \code{clone} is \code{TRUE}, a deep copy is made
if the incoming object is a \code{Graph} to ensure the resulting
object is a different reference from the incoming object.
}
\seealso{
Other Graph operators: \code{\link{\%>>\%}},
  \code{\link{as_pipeop}}, \code{\link{assert_graph}},
  \code{\link{assert_pipeop}}, \code{\link{greplicate}},
  \code{\link{gunion}}
}
\concept{Graph operators}
