% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_graph.R
\name{assert_graph}
\alias{assert_graph}
\title{Assertion for mlr3pipeline Graph}
\usage{
assert_graph(x, coerce = FALSE, deep_copy = FALSE)
}
\arguments{
\item{x}{(\code{any}) \cr
Object to check.}

\item{coerce}{(\code{logical(1)}) \cr
Whether to return a coerced object if possible instead
of throwing an error.}

\item{deep_copy}{(\code{logical(1)}) \cr
Whether to return a deep copy.}
}
\value{
\code{\link{Graph}} \code{x} or a deep clone of it.
}
\description{
Function that checks that a given object is a \code{Graph} and
throws an error if not.

If \code{coerce} is \code{TRUE}, the object is turned into a \code{Graph}
if possible before checking and the coerced object is
returned. If \code{deep_copy} is \code{TRUE}, a deep copy is made
before returning.
}
\seealso{
Other Graph operators: \code{\link{\%>>\%}},
  \code{\link{assert_pipeop}}, \code{\link{greplicate}},
  \code{\link{gunion}}
}
\concept{Graph operators}
