% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{bagging_pipeline}
\alias{bagging_pipeline}
\title{Create a bagging learner}
\usage{
bagging_pipeline(graph, iterations = 10, averager = NULL)
}
\arguments{
\item{`graph`}{\link[=`PipeOp`|`Graph`]{\code{PipeOp}|\code{Graph}} \cr
A learner to create a robustifying pipeline for. Optional, if omitted,
a more conservative pipeline is built.}

\item{`iterations`}{\code{\link{integer}} \cr
Number of bagging iterations. Defaults to 10.}

\item{`averager`}{\link[=`PipeOp`|`Graph`]{\code{PipeOp}|\code{Graph}} \cr}
}
\value{
Returns a \code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} that performs bagging on the inputted Graph
}
\examples{
lrn_po = po("learner", lrn("regr.rpart"))
task = mlr_tasks$get("boston_housing")
gr = bagging_pipeline(lrn_po, 3, averager = po("regravg"))
resample(task, GraphLearner$new(gr), rsmp("holdout"))
}
