% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNewTarget.R
\name{convert_task}
\alias{convert_task}
\title{Convert a task from one type to another.
Requires for the task type to be in \code{mlr_reflections$task_types}.}
\usage{
convert_task(intask, new_type, new_target = NULL)
}
\description{
Convert a task from one type to another.
Requires for the task type to be in \code{mlr_reflections$task_types}.
}
