% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dictionary_sugar_inc_get}
\alias{dictionary_sugar_inc_get}
\title{A Quick Way to Initialize Objects from Dictionaries with Incremented ID}
\usage{
dictionary_sugar_inc_get(dict, .key, ...)
}
\value{
An element from the dictionary.
}
\description{
Covenience function to obtain dictionary elements with incremented ids for easier avoidance
of ID clashes. This should only be used on dictionaries that ensure that the ids don't end with
the pattern \verb{_<n>}. Where \code{n} is any number.
}
\examples{
dictionary_sugar_get("pca", id = "pca_1")
# is the same as
dictionary_sugar_get_inc("pca_1")

}
