% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO_OP.R
\name{filter_noop}
\alias{filter_noop}
\title{Remove NO_OPs from a List}
\usage{
filter_noop(x)
}
\arguments{
\item{x}{\code{list} \cr
List to filter.}
}
\value{
\code{list}: The input list, with all \code{NO_OP} elements removed.
}
\description{
Remove all \code{\link{NO_OP}} elements from a \code{list}.
}
\seealso{
Other Path Branching: \code{\link{NO_OP}},
  \code{\link{is_noop}}, \code{\link{mlr_pipeops_branch}},
  \code{\link{mlr_pipeops_unbranch}}
}
\concept{Path Branching}
