% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\docType{data}
\name{GraphLearner}
\alias{GraphLearner}
\title{GraphLearner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
GraphLearner
}
\description{
GraphLearner
}
\keyword{datasets}
