% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphNode.R
\docType{data}
\name{GraphNode}
\alias{GraphNode}
\title{GraphNode}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GraphNode
}
\description{
The \code{GraphNode} class implements behaviors required to define the
connections between the \code{PipeOps}. The set of interconnected
GraphNodes creates the pipeline graph.
}
\section{Public Members / Active Bindings}{

\itemize{
\item \code{graph}                :: \link{Graph}
Graph that this node is a member of.
\item \code{pipeop}               :: \link{PipeOp}
Contained PipeOp.
\item \code{prev_node_channels}   :: list of \link{NodeChannel}
Identifies the channels from previous nodes that we are connected to.
Mutation can be used to change connections.
\item \code{next_node_channels}   :: list of \link{NodeChannel}
Identifies the channels from successor nodes that we are connected to.
Mutation can be used to change connections.
\item \code{next_nodes}           :: list of \link{GraphNode}
Successor nodes, indexed by channel_id.
\item \code{prev_nodes}           :: list of \link{GraphNode}
Predecessor nodes, indexed by channel_id.
\item \code{in_channels}          :: list of \link{GraphChannel}
Cached identifiers of incoming nodes, indexed by channel id.
\item \code{out_channels}         :: list of \link{GraphChannel}
Cached identifiers of outgoing nodes, indexed by channel id.
\item \code{intype}               :: \link{list}
Identifies types of incoming data, see PipeOp$intype; indexed by channel_id.
\item \code{outtype}              :: \link{list}
Identifies types of outgoing data, see PipeOp$outtype; indexed by channel_id.
\item \code{input_complete}       :: \link{logical(1)}
Whether all incoming channels are connected to other nodes
\item \code{output_complete}      :: \link{logical(1)}
Whether all outgoing channels are connected to other nodes
\item editlock              ????????????????
}
}

\keyword{datasets}
