% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChoice.R
\name{grultiplex}
\alias{grultiplex}
\title{grultiplex}
\usage{
grultiplex(..., .graphs = NULL, .id = "",
  .prefix.gunion.names = FALSE)
}
\arguments{
\item{`...`:}{Multiple graphs, possibly named. They all must have exactly one in_channel and
exactly one out_channel. If any of the arguments are named, then all must have
unique names.}

\item{`.graphs`}{(\link{list of Graph}):
Additionally to the graphs given in \code{...}, a (posibly named) list of graphs can
be given. Default is \code{NULL}.}

\item{`.id`}{(\link{character(1)}):
Optional id prefix to prepend to \code{\link{PipeOpChoice}} and \code{\link{PipeOpUnchoice}} id. Their
resulting IDs will be \code{"choice"} and \code{"unchoice"}, prefixed by \code{.id}. Default is \code{""}.}

\item{`.prefix.gunion.names`}{(\link{logical(1)}):
Whether to add prefixes to graph IDs when performing gunion. Can be helpful to
avoid ID clashes in resulting graph. Default \code{FALSE}.}
}
\description{
Create a multiplexed graph.
}
\examples{
grultiplex(pca = PipeOpPCA$new(), nop = PipeOpNULL$new())
# gives the same as
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpChoice$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnchoice$new(choices)
}
