% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{is.Multiplicity}
\alias{is.Multiplicity}
\title{Check if an object is a Multiplicity}
\usage{
is.Multiplicity(x)
}
\arguments{
\item{x}{(\code{any}) \cr
Object to check.}
}
\value{
\code{logical(1)}
}
\description{
Check if an object is a \code{\link{Multiplicity}}.
}
