% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO_OP.R
\name{is_noop}
\alias{is_noop}
\title{Test for NO_OP}
\usage{
is_noop(x)
}
\arguments{
\item{x}{\code{any} \cr
Object to test.}
}
\value{
\code{logical(1)}: Whether \code{x} is a \code{NO_OP}.
}
\description{
Test whether a given object is a \code{\link{NO_OP}}.
}
\seealso{
Other Path Branching: \code{\link{NO_OP}},
  \code{\link{filter_noop}},
  \code{\link{mlr_pipeops_branch}},
  \code{\link{mlr_pipeops_unbranch}}
}
\concept{Path Branching}
