% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpLearnerRepresentation.R
\name{learner_repr}
\alias{learner_repr}
\alias{learner_repr.xgb.Booster}
\alias{learner_repr.rpart}
\title{Extract a representation from a learner}
\usage{
learner_repr(x, task, ...)

\method{learner_repr}{xgb.Booster}(x, task, ...)

\method{learner_repr}{rpart}(x, task, ...)
}
\arguments{
\item{x}{\code{\link{Learner}} \cr
A trained learner (in the pipeop's state)}

\item{task}{\code{\link{Task}} \cr
A task, the learner was trained on.}

\item{...}{\code{\link{any}} \cr
Further arguments}
}
\value{
\code{\link{data.table}} containing extracted features.
}
\description{
Extract a representation from a learner
}
\details{
for ' xgb.Booster' (xgboost), internally uses xgboost::xgb.create.features.

for 'rpart',  treeClust::rpart.predict.leaves is used.
}
