% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerWeightedAverage.R
\docType{data}
\name{LearnerClassifWeightedAverage}
\alias{LearnerClassifWeightedAverage}
\alias{mlr_learners_classif.weightedaverage}
\title{Weighted Average of Features}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifWeightedAverage
}
\description{
Computes a weighted average of the features.
If \code{weights.method} is "manual", the average is computed over weights provided by the user, defaulting
Predictions are averaged using \code{weights} (in order of appearance in the data); \code{weights} defaults to equal
weights for each feature.

For \code{weights.method}: "nloptr", nonlinear optimization from the package "nloptr" is used to optimize weights
for a measure provided in \code{measure} (defaults to \code{classif.acc}).

Several nonlinear optimization methods from \code{nloptr} can be chosen.
See \code{nloptr::nloptr.print.options()} for a list of possible options.
This can be set via the \code{algorithm} parameter.
}
\section{Methods}{

\itemize{
\item \code{LearnerClassifWeightedAverage$new(), id = "wtaverage")} \cr
(\code{chr}) -> \code{self} \cr
Constructor.
}
}

\concept{LearnerClassif}
\keyword{datasets}
