% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\name{marshal_graph_learner}
\alias{marshal_graph_learner}
\alias{marshal_model.graph_learner_model}
\title{(Un-)Marshal GraphLearner Model}
\usage{
\method{marshal_model}{graph_learner_model}(model, inplace = FALSE, ...)
}
\arguments{
\item{model}{(model of \code{\link{GraphLearner}})\cr
The model to be marshaled.}

\item{inplace}{(\code{logical(1)})\cr
Whether to marshal in-place.
If \code{FALSE} (default), all R6-objects are cloned.}

\item{...}{(any)\cr
Currently unused.}
}
\description{
(Un-) marshal the model of a \code{\link{GraphLearner}}.
}
