% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{mlr_graphs_branching}
\alias{mlr_graphs_branching}
\alias{branch_pipeline}
\title{Create a branch - unbranch pipeline}
\usage{
branch_pipeline(graphs, add_deps = TRUE)
}
\arguments{
\item{graphs}{List of \code{\link{PipeOp}} | \code{\link{Graph}} \cr
A (possibly named) list of \code{\link{PipeOp}}s or \code{\link{Graph}}s.}

\item{add_deps}{\code{\link{logical}} \cr
Whether parameter dependencies should be added. Default to  TRUE`.}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} that provides branching and unbranching
(using \code{\link{PipeOpBranch}} and \code{\link{PipeOpBranch}})
for a list of supplied \code{\link{PipeOp}}s or \code{\link{Graph}}s.
Takes either the names of the list elements, or the graph's ids
as 'option' for \code{\link{PipeOpBranch}}.
}
\examples{
library(mlr3)
library(mlr3misc)
lrns = map(list(lrn("classif.rpart"), lrn("classif.featureless")), po)
task = mlr_tasks$get("boston_housing")
gr = branch_pipeline(lrns)
# change parameters
# gr$param_set$values$branch.selection = "classif.featureless"
}
