% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpBackupLearner.R
\docType{data}
\name{mlr_pipeop_backup_learner}
\alias{mlr_pipeop_backup_learner}
\alias{PipeOpBackupLearner}
\title{PipeOpBackupLearner}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpBackupLearner
}
\description{
Uses an \code{\link[mlr3:Learner]{mlr3::Learner}} to repair predictions done by
a previous \code{\link{PipeOpLearner}}.

Inherits the \code{$param_set} and \code{$values} from the \code{Learner} it is constructed from.
}
\section{Public Members / Active Bindings}{

\itemize{
\item \code{learner}  :: \code{\link{Learner}} \cr
Learner to use for backup prediction. Defaults to
\code{\link[mlr3:mlr_learners_classif.featureless]{classif.featureless}}.
}
}

\section{Methods}{

\itemize{
\item \code{PipeOpLearner$new(learner = mlr_learners$get("classif.featureless", id = "backuplearner", predict_type = "prob"), id = "backuplearner")} \cr
(\code{\link{Learner}}, \code{character(1)}) -> \code{self} \cr
Constructor. The given learner will be used for backup prediction.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}

Other Meta PipeOps: \code{\link{mlr_pipeop_learner_cv}},
  \code{\link{mlr_pipeop_learner}}
}
\concept{Meta PipeOps}
\concept{PipeOps}
\keyword{datasets}
