% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpBalanceSample.R
\docType{data}
\name{mlr_pipeop_balancesample}
\alias{mlr_pipeop_balancesample}
\alias{PipeOpBalanceSample}
\title{PipeOpBalanceSample}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpBalanceSample
}
\description{
Both undersamples (filters) a \code{\link{Task}} to keep only a fraction of the rows with majority class,
as well as oversamples (repeats datapoints) minorit-class rows.

Possible parameters are \code{$ratio}, \code{$reference}, and \code{$adjust}.
}
\section{Parameter Set}{

\itemize{
\item \code{ratio} :: \code{numeric(1)} \cr
Ratio of number of rows of classes to keep, relative
to the \code{$reference} value.
\item \code{reference} :: \code{numeric(1)} \cr
What the \code{$ratio} value is measured against. Can be \code{"all"} (default, mean instance count of
all classes), \code{"major"} (instance count of class with most instances), \code{"minor"}
(instance count of class with fewest instances), \code{"nonmajor"} (average instance
count of all classes except the major one), \code{"nonminor"} (average instance count
of all classes except the minor one), and \code{"one"} (\code{$ratio} determines the number of
instances to have, per class).
\item \code{adjust} :: \code{numeric(1)} \cr
Which classes to up / downsample. Can be \code{"all"} (default, up and downsample all to match required
instance count), \code{"major"}, \code{"minor"}, \code{"nonmajor"}, \code{"nonminor"} (see respective values
for \code{$reference}), \code{"upsample"} (only upsample), and \code{"downsample"}.
\item \code{shuffle} :: \code{logical(1)} \cr
Whether to shuffle the result. Otherwise, the resulting task will have the original items that
were not removed in downsampling in-order, followed by all newly sampled items ordered by target class.
Default is \code{TRUE}.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
