% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChunk.R
\docType{data}
\name{mlr_pipeop_chunk}
\alias{mlr_pipeop_chunk}
\alias{PipeOpChunk}
\title{PipeOpChunk}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpChunk
}
\description{
Chunks its input into \code{outnum} chunks.
Returns a list of \code{\link{Task}}s during training, and
simply passes on the input during prediction.
}
\section{Methods}{

\itemize{
\item \code{PipeOpChunk$new(outnum, id = "chunk")} \cr
(\code{integer(1)}, \code{character(1)}) -> \code{self} \cr
Constructor. \code{outnum} gives the number of output
channels / chunks that are created.
}
}

\section{Parameter Set}{

\itemize{
\item \code{shuffle} :: \code{logical(1)} \cr
Should the data be shuffled before chunking? Default \code{TRUE}
\item \code{stratify} :: \code{logical(1)} \cr
Should the subsamples be stratified. Default \code{FALSE}.
}
}

\examples{
op = PipeOpChunk$new(5)
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_nloptmajorityvote}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeop_undersample}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
