% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDownsample.R
\docType{data}
\name{mlr_pipeop_downsample}
\alias{mlr_pipeop_downsample}
\alias{PipeOpDownsample}
\title{PipeOpDownsample}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpDownsample
}
\description{
Subsamples a \code{\link{Task}} to use only a fraction of the rows.
}
\section{Parameter Set}{

\itemize{
\item \code{frac} :: \code{numeric(1)} \cr
Fraction of rows in the task to keep. Default 1.
\item \code{stratify} :: \code{logical(1)} \cr
Should the subsamples be stratified by target? Default \code{FALSE}.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
