% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFeatureUnion.R
\docType{data}
\name{mlr_pipeop_featureunion}
\alias{mlr_pipeop_featureunion}
\alias{PipeOpFeatureUnion}
\title{PipeOpFeatureUnion}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpFeatureUnion
}
\description{
Aggregates features from all input tasks by cbinding them together
into a \code{\link{data.table}}.
\code{\link{DataBackend}} primary keys and \code{\link{Task}} targets have to be equal across each
\code{Task}. Only one target columns is kept.
}
\section{Methods}{

\itemize{
\item \code{PipeOpFeatureUnion$new(innum, id = "featureunion")} \cr
(\code{numeric(1)}, \code{character(1)}) -> \code{self} \cr
Constructor. \code{innum} determines the number of input channels.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
