% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpImpute.R
\docType{data}
\name{mlr_pipeop_impute}
\alias{mlr_pipeop_impute}
\alias{PipeOpImpute}
\title{PipeOpImpute}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PipeOpImpute
}
\description{
Impute missing values with varying methods.
\itemize{
\item \code{numeric} or \code{integer} features are imputed by \code{method_num}.
\item \code{factor}, \code{ordered}, and \code{character} features are imported by \code{method_fct}.
\item \code{logical} features are always imputed by sampling from the training column.
}
}
\examples{
task = mlr3::mlr_tasks$get("pima")
sum(complete.cases(task$data()))

po = mlr_pipeops$get("impute")
new_task = po$train(list(task = task))[[1]]
sum(complete.cases(new_task$data()))
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
