% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpLearner.R
\docType{data}
\name{mlr_pipeop_learner}
\alias{mlr_pipeop_learner}
\alias{PipeOpLearner}
\title{PipeOpLearner}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpLearner
}
\description{
Wraps an \code{\link[mlr3:Learner]{mlr3::Learner}} into a \code{\link{PipeOp}}.
Inherits the \code{$param_set} and \code{$values} from the \code{Learner} it is constructed from.
}
\section{Public Members / Active Bindings}{

\itemize{
\item \code{learner}  :: \code{\link{Learner}} \cr
Learner to use for cross validation / prediction.
}
}

\section{Methods}{

\itemize{
\item \code{PipeOpLearner$new(learner, id = learner$id)} \cr
(\code{\link{Learner}}, \code{character(1)}) -> \code{self} \cr
Constructor. The given learner will be used for crossvalidation.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeop_undersample}},
  \code{\link{mlr_pipeops}}

Other Meta PipeOps: \code{\link{mlr_pipeop_learner_cv}}
}
\concept{Meta PipeOps}
\concept{PipeOps}
\keyword{datasets}
