% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpLearnerCV.R
\docType{data}
\name{mlr_pipeop_learner_cv}
\alias{mlr_pipeop_learner_cv}
\alias{PipeOpLearnerCV}
\title{PipeOpLearnerCV}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpLearnerCV
}
\description{
Wraps an \code{\link[mlr3:Learner]{mlr3::Learner}} into a \code{\link{PipeOp}}.
Returns cross-validated predictions during training phase and stores a model of the
\code{Learner} trained on the whole data in \code{$state} as a new \code{Task} with the predictions;
usually a feature \code{<ID>.response} if the \code{Learner}'s \code{$predict.type} is \code{"response"},
\code{<ID>.prob.<CLASS>} and \code{<ID>.response} for \code{$predict.type} \code{"prob"}, or \code{<ID>.response}
and \code{<ID>.se} for \code{$predict.type} \code{"se"}. \code{<ID>} is here the \code{$id} of the \code{PipeOpLearnerCV}.

Returns this model's prediction during prediction phase, as a new \code{Task} with a single
column.

Inherits the \code{$param_set} and \code{$param_vals} from the \code{Learner} it is constructed from.
}
\section{Public Members / Active Bindings}{

\itemize{
\item \code{learner}  :: \code{\link{Learner}} \cr
Learner to use for cross validation / prediction.
}
}

\section{Methods}{

\itemize{
\item \code{PipeOpLearner$new(learner, id = learner$id)} \cr
(\code{\link{Learner}}, \code{character(1)}) -> \code{self} \cr
Constructor. The given learner will be used for crossvalidation.
}
}

\section{Parameter Set}{

\itemize{
\item \code{resampling} :: \code{character(1)} \cr
Which resampling method do we want to use. Currently only supports 'cv'.
\item \code{folds}     :: \code{numeric(1)} \cr
Number of cross validation folds.
\item \code{...} \cr
The \code{\link{ParamSet}} of \code{$learner} is also made available.
}
}

\seealso{
Other Meta PipeOps: \code{\link{mlr_pipeop_learner}}
}
\concept{Meta PipeOps}
\concept{Pipeops}
\keyword{datasets}
