% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{mlr_pipeop_majorityvote}
\alias{mlr_pipeop_majorityvote}
\alias{PipeOpMajorityVote}
\title{PipeOpMajorityVote}
\format{\link{R6Class} PipeOpMajorityVote}
\usage{
PipeOpMajorityVote
}
\description{
Aggregates over different \code{\link{PredictionDataClassif}}s.
Either computes the mode, if \code{predict_type} is \code{"response"},
or averages probabilities if \code{predict_type} is \code{"prob"}.
Returns a single \code{\link{PredictionDataClassif}}.
Weights can be set by the user, if none are provided, defaults to
equal weights \code{rep(1/innum, innum)} for each prediction.
Used for classification \code{Prediction}s.
Defaults to equal weights for each model.
}
\examples{
op = PipeOpMajorityVote$new(3)
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
