% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpModelAvg.R
\docType{data}
\name{mlr_pipeop_modelavg}
\alias{mlr_pipeop_modelavg}
\alias{PipeOpModelAvg}
\title{PipeOpModelAvg}
\format{\code{\link{R6Class}} inheriting from \code{\link{PipeOpEnsemble}}.}
\usage{
PipeOpModelAvg
}
\description{
Averages its input (a \code{list} of \code{PredictionRegr}).
Only used for regression \code{Prediction}s.
Weights can be set by the user, if none are provided, defaults to
equal weights for each prediction.
Offers a \code{$weights} slot to set/get weights for each learner.
Returns a single \code{PredictionRegr}.
Defaults to equal weights for each model.
}
\examples{
op = PipeOpModelAvg$new()
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
