% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpMutate.R
\docType{data}
\name{mlr_pipeop_mutate}
\alias{mlr_pipeop_mutate}
\alias{PipeOpMutate}
\title{PipeOpMutate}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreprocSimple}}.}
\usage{
PipeOpMutate
}
\description{
Adds features according to expressions.
}
\section{Parameter Set}{

\itemize{
\item \code{mutation} :: named \code{list} \cr
Expressions for new features to create. Can be created with \code{alist()}.
Can contain multiple expressions, that directly reference task features.
\item \code{env} :: \code{environment} \cr
Environment in which expressions are evaluated. This is the enclosing
environment \emph{after} the \code{Task}'s features are considered.
\item \code{delete_originals} :: \code{logical(1)} \cr
Whether to delete original features. If this is \code{FALSE} (the default),
then features may still be overwritten.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_nop}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
