% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{mlr_pipeop_nloptmajorityvote}
\alias{mlr_pipeop_nloptmajorityvote}
\alias{PipeOpNlOptModelAvg}
\alias{PipeOpNlOptMajorityVote}
\title{PipeOpNlOptModelAvg}
\format{\link{R6Class} PipeOpNlOptModelAvg}
\usage{
PipeOpNlOptModelAvg

PipeOpNlOptMajorityVote
}
\description{
Aggregates over different \code{\link{PredictionRegr}}s.
Weights for each learner are learned using (nloptr)\link[nloptr:nloptr]{nloptr::nloptr}.
For help with nloptr see \code{\link[nloptr:nloptr.print.options]{nloptr::nloptr.print.options()}}.
Returns a single \code{\link{PredictionRegr}}.
By default, optimizes \code{\link{MeasureRegrMSE}} and only allows weights between 0 and 1.
Used for regression \code{\link{Prediction}}s.

Aggregates over different \code{\link{PredictionClassif}}s.
Either computes the mode, if \code{predict_type} is \code{"response"},
or averages probabilities if \code{predict_type} is \code{"prob"}.
Weights for each learner are learned using (nloptr)\link[nloptr:nloptr]{nloptr::nloptr}.
For help with nloptr see \code{\link[nloptr:nloptr.print.options]{nloptr::nloptr.print.options()}}.
Returns a single \code{\link{PredictionClassif}}.
As a default, optimizes \code{\link{MeasureClassifMMCE}} and only allows weights between 0 and 1.
Used for classification \code{\link{Prediction}}s.
}
\examples{
op = PipeOpNlOptModelAvg$new(3)
op = PipeOpNlOptMajorityVote$new(3)
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}

Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
