% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpPCA.R
\docType{data}
\name{mlr_pipeop_pca}
\alias{mlr_pipeop_pca}
\alias{PipeOpPCA}
\title{PipeOpPCA}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpPCA
}
\description{
Extracts principle components from data.
See \code{\link[stats:prcomp]{stats::prcomp()}} for details  and parameters.
}
\examples{
# Instantiate PipeOpPCA
op1 = PipeOpPCA$new()
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
