% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpSelect.R
\docType{data}
\name{mlr_pipeop_select}
\alias{mlr_pipeop_select}
\alias{PipeOpSelect}
\title{PipeOpSelect}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreprocSimple}}.}
\usage{
PipeOpSelect
}
\description{
Removes features from \code{Task} depending on a selector function. If \code{invert}
is \code{FALSE} (default) then \code{selector} gives the features to keep, otherwise it gives
the features to drop. See \code{\link{Selector}} for selectors that are provided.
}
\section{Parameter Set}{

\itemize{
\item \code{selector} :: \code{function} \cr
Selector function, takes a \code{Task} as argument and returns a \code{character}
of features to keep. See \code{\link{Selector}} for example functions. Defaults to
\code{selector_all()}.
\item \code{invert}    :: \code{logical(1)} \cr
Invert selection. If this is \code{TRUE}, the features selected by
\code{selector} are \emph{removed} while the ones not selected remain.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}

Other Selectors: \code{\link{Selector}}
}
\concept{PipeOps}
\concept{Selectors}
\keyword{datasets}
