% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpSubsample.R
\docType{data}
\name{mlr_pipeop_subsample}
\alias{mlr_pipeop_subsample}
\alias{PipeOpSubsample}
\title{PipeOpSubsample}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpSubsample
}
\description{
Subsamples a \code{\link{Task}} to use a fraction of the rows.
\code{frac} can be > 1 if \code{replace} is \code{TRUE}.
}
\section{Parameter Set}{

\itemize{
\item \code{frac} :: \code{numeric(1)} \cr
Fraction of rows in the task to keep. Default 1.
\item \code{stratify} :: \code{logical(1)} \cr
Should the subsamples be stratified by target? Default \code{FALSE}.
\item \code{replace} :: \code{logical(1)} \cr
Sample with replacement? Default is \code{FALSE}.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_nop}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
