% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{mlr_pipeop_weightedmodelavg}
\alias{mlr_pipeop_weightedmodelavg}
\alias{PipeOpWeightedModelAvg}
\title{PipeOpWeightedModelAvg}
\format{\code{\link{R6Class}} inheriting from \code{\link{PipeOpEnsemble}}.}
\usage{
PipeOpWeightedModelAvg
}
\description{
Averages its input (a \code{list} of \code{\link{Prediction}}).
Weights can be learned to optimize a specified
measure.
Returns a single \code{\link{Prediction}}.
Used for regression \code{Prediction}s.
}
\examples{
op = PipeOpWeightedModelAvg$new(3)
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_downsample}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeop_undersample}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
