% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{Multiplicity}
\alias{Multiplicity}
\title{Multiplicity}
\usage{
Multiplicity(...)
}
\arguments{
\item{\dots}{\cr
Can be anything.}
}
\value{
\code{\link{Multiplicity}}
}
\description{
Construct an object of the S3 class \code{\link{Multiplicity}}, which is a named or
unnamed lists of data.
}
