% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpBranch.R
\docType{data}
\name{NO_OP}
\alias{NO_OP}
\title{No-Op Sentinel Used for Alternative Branching}
\format{An object of class \code{NO_OP} (inherits from \code{R6}) of length 3.}
\usage{
NO_OP
}
\description{
Special data type for no-ops. Distinct from NULL for easier
and distinction from unintentional NULL returns.
}
\keyword{datasets}
