% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{Optimizer}
\alias{Optimizer}
\title{Optimizer Base class}
\format{
Abstract \code{\link{R6Class}}.
}
\description{
Abstract base class for Optimizers. Exposes a \verb{$optimize} method
that can be used for optimization.
}
\section{Construction}{
\preformatted{Optimizer$new(param_set = ParamSet$new(), param_vals = list(), packages = character(0))
}
\itemize{
\item \code{param_set} :: \code{\link[paradox:ParamSet]{ParamSet}} | \code{list} of \code{expression}\cr
Parameter space description. This should be created by the subclass and given to \code{super$initialize()}.
If this is a \code{\link[paradox:ParamSet]{ParamSet}}, it is used as the \code{\link{PipeOp}}'s \code{\link[paradox:ParamSet]{ParamSet}}
directly. Otherwise it must be a \code{list} of expressions e.g. created by \code{alist()} that evaluate to \code{\link[paradox:ParamSet]{ParamSet}}s.
These \code{\link[paradox:ParamSet]{ParamSet}} are combined using a \code{\link[paradox:ParamSetCollection]{ParamSetCollection}}.
Default: ParamSet$new().
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings given in \code{param_set}. The
subclass should have its own \code{param_vals} parameter and pass it on to \code{super$initialize()}. Default \code{list()}.
\item packages :: \code{character}\cr
Set of all required packages for the \code{\link{PipeOp}}'s \verb{$train} and \verb{$predict} methods. See \verb{$packages} slot.
Default is \code{character(0)}.
}
}

\section{Fields}{

\itemize{
\item \code{packages} :: \code{character}\cr
Packages required for the \code{\link{Optimizer}}. Functions that are not in base R should still be called using \code{::}.
\item \code{param_set} :: \code{\link[paradox:ParamSet]{ParamSet}}\cr
Parameters and parameter constraints. Parameter values that influence the functioning of \verb{$optimize}.
}
}

\section{Methods}{

\itemize{
\item \code{optimize(objfun, init_weights = NULL, ...)}\cr
(\code{function}, \code{numeric}, \code{any}) -> \code{numeric}\cr
Runs the optimization and returns optimal values.
\code{...} is passed on to the optimizer.
Some optimizers might require \code{init_weights}, the starting values for the optimization.
}
}

\seealso{
Other Optimizer: 
\code{\link{OptimizerGenSA}},
\code{\link{OptimizerNloptr}}
}
\concept{Optimizer}
