% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{OptimizerNloptr}
\alias{OptimizerNloptr}
\title{Optimizer from nloptr}
\format{Abstract \code{\link{R6Class}}.}
\description{
Non-linear optimization using \code{nloptr::nloptr}, called with parameters specified in
param_vals`.
}
\section{Construction}{
\preformatted{Optimizer$new(param_vals = list())
}
\itemize{
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings given in \code{param_set}. The
subclass should have its own \code{param_vals} parameter and pass it on to \code{super$initialize()}. Default \code{list()}.
}
}

\section{Fields}{

\itemize{
\item \code{packages} :: \code{character}\cr
Packages required for the \code{\link{Optimizer}}. Functions that are not in base R should still be called using \code{::}.
\item \code{param_set} :: \code{\link[paradox:ParamSet]{ParamSet}}\cr
Parameters and parameter constraints. Parameter values that influence the functioning of \verb{$optimize}.
}
}

\section{Methods}{

\itemize{
\item \code{optimize}, see \code{\link{Optimizer}} for details.
}
}

\seealso{
Other Optimizer: 
\code{\link{OptimizerGenSA}},
\code{\link{Optimizer}}
}
\concept{Optimizer}
