% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{pipeline_greplicate}
\alias{pipeline_greplicate}
\title{Create Disjoint Graph Union of Copies of a Graph}
\usage{
pipeline_greplicate(graph, n)
}
\arguments{
\item{graph}{\code{\link{Graph}} \cr
Graph to replicate.}

\item{n}{\code{integer(1)}
Number of copies to create.}
}
\value{
\code{\link{Graph}} containing \code{n} copies of input \code{graph}.
}
\description{
Create a new \code{\link{Graph}} containing \code{n} copies of the input \code{\link{Graph}} / \code{\link{PipeOp}}. To avoid ID
collisions, PipeOp IDs are suffixed with \verb{_i} where \code{i} ranges from 1 to \code{n}.
}
\examples{
library("mlr3")

po_pca = po("pca")
pipeline_greplicate(po_pca, n = 2)
}
