% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChoice.R
\docType{data}
\name{PipeOpChoice}
\alias{PipeOpChoice}
\title{PipeOpChoice}
\format{\link{R6Class} PipeOpChoice}
\usage{
PipeOpChoice
}
\description{
This pipeop is used for multiplexing between different possible paths.
}
\section{Methods}{

\itemize{
\item \code{new(options = 1, id = "choice")} \cr
(\code{integer(1)} | \code{character}), \code{character(1)} -> \code{\link{PipeOpChoice}}
}
}

\section{Parameters}{

\itemize{
\item \code{selection}: integer or discrete
}
}

\section{Details}{

Creates a PipeOp with multiple output channels that can be used to
create a Graph network with alternative paths. If \code{options} is an \code{integer(1)},
it determines the number of out-paths and \code{selection} is an integer parameter
choosing between these paths. If \code{options} is a \code{character}, then \code{length(options)}
out channels are created, each named according to \code{options}.

To create a usable graph, the branching paths need to be brought together
using \code{\link{PipeOpUnchoice}}.
}

\examples{
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpChoice$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnchoice$new(choices)

}
\seealso{
Other PipeOp: \code{\link{PipeOpDT}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnchoice}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
