% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChunk.R
\docType{data}
\name{PipeOpChunk}
\alias{PipeOpChunk}
\title{PipeOpChunk}
\format{\link{R6Class} PipeOpChunk}
\usage{
PipeOpChunk
}
\description{
Chunks its input into \code{outnum} chunks.
Returns a list of \link[mlr3:Task]{mlr3::Task}.
During predict simply passes on the input.
}
\section{Usage}{

Inherits from \link{PipeOpChunk}
\itemize{
\item \code{f = PipeOpChunk$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpCopy}
}
}

\section{Details}{

\itemize{
\item \code{outnum}: \code{integer(1)} Number of times the input is copied.
\item \code{shuffle}: \code{logical(1)} Should the data be shuffled before chunking?
\item \code{stratify}: \code{logical(1)} Should the subsamples be stratified.
}
}

\examples{
op = PipeOpChunk$new(5)
}
\concept{PipeOp, PipeOpBroadcast, PipeOpChunk}
\keyword{datasets}
