% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpCopy.R
\docType{data}
\name{PipeOpCopy}
\alias{PipeOpCopy}
\title{PipeOpCopy}
\format{\link{R6Class} PipeOpCopy}
\usage{
PipeOpCopy
}
\description{
Copies its input \code{outnum} times.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = PipeOpCopy$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpCopy}
}
}

\section{Details}{

\itemize{
\item \code{outnum}: \code{integer(1)} Number of times the input is copied.
}
}

\seealso{
Other PipeOp, PipeOpBroadcast: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}
}
\concept{PipeOp, PipeOpBroadcast}
\keyword{datasets}
