% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDownsample.R
\docType{data}
\name{PipeOpDownsample}
\alias{PipeOpDownsample}
\title{PipeOpDownsample}
\format{\link{R6Class} PipeOpDownsample}
\usage{
PipeOpDownsample
}
\description{
Subsamples a \link{Task} to include only a fraction of the rows.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = pipeOpDownsample$new(id)} \cr
}
}

\section{Details}{

\itemize{
\item \code{perc}: \code{numeric(1)} Percentage of rows in the task to keep.
\item \code{stratify}: \code{logical(1)} Should the subsamples be stratified.
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpDT}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
