% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpFeatureUnion.R
\docType{data}
\name{PipeOpFeatureUnion}
\alias{PipeOpFeatureUnion}
\title{PipeOpFeatureUnion}
\format{\link{R6Class} PipeOpFeatureUnion}
\usage{
PipeOpFeatureUnion
}
\description{
Aggregates features from all input tasks by cbinding them together
into a \link{data.table}.
\link{DataBackend} primary keys and \link{Task} targets have to be equal across each
task. Only one target is kept.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = pipeOpFeatureUnion$new(id)} \cr
\code{character(1)} -> \link{PipeOpFeatureUnion}
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnbranch}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
