% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpLearner.R
\docType{data}
\name{PipeOpLearner}
\alias{PipeOpLearner}
\title{PipeOpLearner}
\format{\link{R6Class} PipeOpLearner}
\usage{
PipeOpLearner
}
\description{
Wraps a \link[mlr3:Learner]{mlr3::Learner} into a \link{PipeOp}.
Inherits the \code{param_set} from the \link[mlr3:Learner]{mlr3::Learner} it is constructed from.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = PipeOpLearner$new(outnum, id)} \cr
\code{[Learner]} -> \link{PipeOpLearner}
}
}

\concept{PipeOp, PipeOpLearner}
\keyword{datasets}
