% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\name{PipeOpMajorityVote}
\alias{PipeOpMajorityVote}
\title{PipeOpMajorityVote}
\format{\link{R6Class} PipeOpMajorityVote}
\description{
Computes the mode over different predictions for each row_id.
}
\section{Usage}{

Inherits from \link{PipeOpMajorityVote}
\itemize{
\item \code{f = PipeOpMajorityVote$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpMajorityVote}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\examples{
op = PipeOpMajorityVote$new(3)
}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}

Other PipeOpAggregate: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpModelAvg}},
  \code{\link{PipeOpUnbranch}}

Other PipeOpEnsemble: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpModelAvg}}
}
\concept{PipeOp}
\concept{PipeOpAggregate}
\concept{PipeOpEnsemble}
