% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{PipeOpModelAvg}
\alias{PipeOpModelAvg}
\title{PipeOpModelAvg}
\format{\link{R6Class} PipeOpModelAvg}
\usage{
PipeOpModelAvg
}
\description{
Averages its input (a list of \link[mlr3:Prediction]{mlr3::Prediction}).
Returns a \link[mlr3:Prediction]{mlr3::Prediction}.
}
\section{Usage}{

Inherits from \link{PipeOpEnsemble}
\itemize{
\item \code{f = PipeOpModelAvg$new(innum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpModelAvg}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\examples{
op = PipeOpModelAvg$new(3)
}
\seealso{
Other PipeOp, PipeOpAggregate, PipeOpEnsemble: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpMajorityVote}}
}
\concept{PipeOp, PipeOpAggregate, PipeOpEnsemble}
\keyword{datasets}
