% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNewTarget.R
\docType{data}
\name{PipeOpNewTarget}
\alias{PipeOpNewTarget}
\title{PipeOpNewTarget}
\format{Abstract \code{\link{R6Class}} inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpNewTarget
}
\description{
Substitutes the current target with a new target.
The new target must exist as a column in the Task.
The previous target is set to row_role: `unused'.
}
\section{Public Members / Active Bindings}{

\itemize{
\item \code{new_target} :: \code{character} \cr
New target variable name for the task. The previous target variable is set to 'unused'.
\item \code{new_task_type} :: \code{NULL} | \code{character} \cr
New task type for the resulting task. Defaults to 'classif' if \code{new_target}is \code{character} | \code{factor},
else \code{regr}.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeops_branch}},
  \code{\link{mlr_pipeops_chunk}},
  \code{\link{mlr_pipeops_classbalancing}},
  \code{\link{mlr_pipeops_classifavg}},
  \code{\link{mlr_pipeops_colapply}},
  \code{\link{mlr_pipeops_copy}},
  \code{\link{mlr_pipeops_encode}},
  \code{\link{mlr_pipeops_featureunion}},
  \code{\link{mlr_pipeops_filter}},
  \code{\link{mlr_pipeops_impute}},
  \code{\link{mlr_pipeops_learner}},
  \code{\link{mlr_pipeops_mutate}},
  \code{\link{mlr_pipeops_nop}},
  \code{\link{mlr_pipeops_pca}},
  \code{\link{mlr_pipeops_regravg}},
  \code{\link{mlr_pipeops_scale}},
  \code{\link{mlr_pipeops_select}},
  \code{\link{mlr_pipeops_subsample}},
  \code{\link{mlr_pipeops_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
