% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpPCA.R
\docType{data}
\name{PipeOpPCA}
\alias{PipeOpPCA}
\title{PipeOpPCA}
\format{\link{R6Class} PipeOpPCA}
\usage{
PipeOpPCA
}
\description{
Extracts principle components from data.
See \link[stats:prcomp]{stats::prcomp} for details  and parameters.
}
\section{Usage}{

Inherits from \link{PipeOpDT}
\itemize{
\item \code{f = pipeOpPCA$new(id)} \cr
\code{character(1)} -> \link{PipeOpPCA}
}
}

\examples{
# Instantiate PipeOpPCA
op1 = PipeOpPCA$new()
}
\seealso{
Other PipeOp: \code{\link{PipeOpDT}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
