% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpScale.R
\docType{data}
\name{PipeOpScale}
\alias{PipeOpScale}
\title{PipeOpScale}
\format{\link{R6Class} PipeOpScale}
\usage{
PipeOpScale
}
\description{
Scales the data to mean = 0 and standard deviation 1.
See \link[base:scale]{base::scale} for details and parameters.
}
\section{Usage}{

Inherits from \link{PipeOpDT}
\itemize{
\item \code{f = pipeOpDT$new(id)} \cr
\code{character(1)} -> \link{PipeOpDT}
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpChoice}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpUnchoice}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
