% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChoice.R
\docType{data}
\name{PipeOpUnchoice}
\alias{PipeOpUnchoice}
\title{PipeOpUnchoice}
\format{\link{R6Class} PipeOpUnchoice}
\usage{
PipeOpUnchoice
}
\description{
Used to bring together different paths created by \code{\link{PipeOpChoice}}.
}
\examples{
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpChoice$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnchoice$new(choices)

}
\seealso{
Other PipeOp: \code{\link{PipeOpChoice}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
