% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecheck.R
\name{register_autoconvert_function}
\alias{register_autoconvert_function}
\title{Add Autoconvert Function to Conversion Register}
\usage{
register_autoconvert_function(cls, fun, packages = character(0))
}
\arguments{
\item{cls}{\code{character(1)} The class that \code{fun} converts to.}

\item{fun}{\code{function} The conversion function. Must take one
argument and return an object of class \code{cls}, or possibly
a sub-class as recognized by \code{are_types_compatible()}.}

\item{packages}{\code{character} The packages required to be loaded for fun to operate.}
}
\description{
Add functions that perform conversion to a desired class.
}
\seealso{
Other class hierarchy operations: \code{\link{add_class_hierarchy_cache}},
  \code{\link{reset_autoconvert_register}},
  \code{\link{reset_class_hierarchy_cache}}
}
\concept{class hierarchy operations}
