% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecheck.R
\name{reset_autoconvert_register}
\alias{reset_autoconvert_register}
\title{Reset Autoconvert Register}
\usage{
reset_autoconvert_register()
}
\description{
Reset autoconvert register.
}
\seealso{
Other class hierarchy operations: \code{\link{add_class_hierarchy_cache}},
  \code{\link{register_autoconvert_function}},
  \code{\link{reset_class_hierarchy_cache}}
}
\concept{class hierarchy operations}
