% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typecheck.R
\name{reset_class_hierarchy_cache}
\alias{reset_class_hierarchy_cache}
\title{Reset the Class Hierarchy Cache}
\usage{
reset_class_hierarchy_cache()
}
\description{
Reset the class hierarchy cache to factory default.
}
\seealso{
Other class hierarchy operations: \code{\link{add_class_hierarchy_cache}},
  \code{\link{register_autoconvert_function}},
  \code{\link{reset_autoconvert_register}}
}
\concept{class hierarchy operations}
