% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{robustify_pipeline}
\alias{robustify_pipeline}
\title{Robustify a learner}
\usage{
robustify_pipeline(
  task,
  learner = NULL,
  scaling = TRUE,
  max_cardinality = 1000
)
}
\arguments{
\item{task}{\code{\link{Task}} \cr
A task to create a robustifying pipeline for.}

\item{learner}{\code{\link{Learner}} \cr
A learner to create a robustifying pipeline for. Optional, if omitted,
a more conservative pipeline is built.}

\item{scaling}{\code{\link{logical}} \cr
Should the data be scaled? Defaults to true.}

\item{max_cardinality}{\code{\link{integer}} \cr
Maximum number of factor levels allowed. See above. Default: 1000.}
}
\value{
Returns a \code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} that can be used to robustify any subsequent learner.
Performs the following steps:
\itemize{
\item Drops empty factor levels using \code{\link{PipeOpFixFactors}}
\item Imputes \code{numeric} features using \code{\link{PipeOpImputeHist}} and \code{\link{PipeOpMissInd}}
\item Imputes \code{factor} features using \code{\link{PipeOpImputeNewlvl}}
\item Encodes \code{factors} using \code{one-hot-encoding}. Factors with a cardinality > \code{ḿax_cardinality} are
collapsed using \code{\link{PipeOpCollapseFactors}}.
\item If \code{scaling}, numeric features are scaled to mean 0 and standard deviation 1.
}

The graph is built conservatively, i.e. the function always tries to assure everything works.
If a learner is provided, some steps can be left out, i.e. if the learner can deal with
factor variables, no encoding is performed.
}
\examples{
lrn = lrn("regr.rpart")
task = mlr_tasks$get("boston_housing")
gr = robustify_pipeline(task, lrn) \%>>\% po("learner", lrn)
resample(task,GraphLearner$new(gr), rsmp("holdout"))
}
