% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selector.R
\name{Selector}
\alias{Selector}
\alias{selector_all}
\alias{selector_type}
\alias{selector_grep}
\alias{selector_name}
\title{Selector Functions}
\usage{
selector_all()

selector_type(types)

selector_grep(pattern, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE)

selector_name(feature_names, assert_present = FALSE)
}
\arguments{
\item{types}{(\code{character}) \cr
Type of feature to select}

\item{pattern}{(\code{character(1)}) \cr
grep pattern}

\item{ignore.case}{(\code{logical(1)}) \cr
ignore case}

\item{perl}{(\code{logical(1)}) \cr
perl regex}

\item{fixed}{(\code{logical(1)}) \cr
fixed pattern instead of regex}

\item{feature_names}{(\code{character}) \cr
select features by exact name match}

\item{assert_present}{(\code{logical(1)}) \cr
error if \code{feature_names} are not all present}
}
\description{
Selector functions.
}
\seealso{
Other Selectors: \code{\link{mlr_pipeop_select}}
}
\concept{Selectors}
