% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\name{set_inner_tuning.GraphLearner}
\alias{set_inner_tuning.GraphLearner}
\title{Set Inner Tuning of a GraphLearner}
\usage{
\method{set_inner_tuning}{GraphLearner}(.learner, .disable = FALSE, validate = NA, args = NULL, ...)
}
\arguments{
\item{validate}{(\code{numeric(1)}, \code{"inner_valid"}, or \code{NULL})\cr
How to set the \verb{$validate} field of the learner.}

\item{args}{(named \code{list()})\cr
Names are ids of the \code{\link{GraphLearner}}'s \code{PipeOps} and values are lists containing arguments passed to the
respective wrapped \code{\link{Learner}}.
By default, the values \code{.disable} and \code{validate} are used, but can be overwritten on a per-pipeop basis.

When enabling, the inner tuning of the \verb{$base_learner()} is enabled by default.
When disabling, all inner tuning is disable by default.}
}
\description{
First, all values specified by \code{...} are
All \code{\link{PipeOpLearner}} and \code{\link{PipeOpLearnerCV}}
}
