% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_graph.R
\name{assert_graph}
\alias{assert_graph}
\title{Assertion for mlr3pipeline Graph}
\usage{
assert_graph(x)
}
\arguments{
\item{x}{(\code{any}) \cr
Object to check.}
}
\value{
\code{\link{Graph}} \code{invisible(x)}
}
\description{
Function that checks that a given object is a \code{Graph} and
throws an error if not.
}
\seealso{
Other Graph operators: 
\code{\link{\%>>\%}()},
\code{\link{as_graph}()},
\code{\link{as_pipeop}()},
\code{\link{assert_pipeop}()},
\code{\link{greplicate}()},
\code{\link{gunion}()},
\code{\link{pipeline_greplicate}()}
}
\concept{Graph operators}
