% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_graph.R
\name{assert_pipeop}
\alias{assert_pipeop}
\title{Assertion for mlr3pipelines PipeOp}
\usage{
assert_pipeop(x)
}
\arguments{
\item{x}{(\code{any}) \cr
Object to check.}
}
\value{
\code{\link{PipeOp}} \code{invisible(x)}
}
\description{
Function that checks that a given object is a \code{PipeOp} and
throws an error if not.
}
\seealso{
Other Graph operators: 
\code{\link{as_graph}()},
\code{\link{as_pipeop}()},
\code{\link{assert_graph}()},
\code{\link{chain_graphs}()},
\code{\link{concat_graphs}()},
\code{\link{greplicate}()},
\code{\link{gunion}()},
\code{\link{pipeline_greplicate}()}
}
\concept{Graph operators}
