% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNewTarget.R
\name{convert_task}
\alias{convert_task}
\title{Convert a task from one type to another.
Requires for the task type to be in \code{mlr_reflections$task_types}.
The new target is set as a new target, while the previous target is added
as a feature.}
\usage{
convert_task(intask, new_type, new_target = NULL)
}
\arguments{
\item{intask}{\code{\link{Task}}\cr
A \code{\link{Task}} to be converted.}

\item{new_type}{\code{character(1)}\cr
The new task type. Must be in  \code{mlr_reflections$task_types}.}

\item{new_target}{\code{character(1)|NULL}\cr
New target to be set, must be a column in the \code{intask} data.
If NULL, no new target is set.}
}
\value{
\code{\link{Task}}
}
\description{
Convert a task from one type to another.
Requires for the task type to be in \code{mlr_reflections$task_types}.
The new target is set as a new target, while the previous target is added
as a feature.
}
