% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNewTarget.R
\name{convert_task}
\alias{convert_task}
\title{Convert a task from its type to another.}
\usage{
convert_task(
  intask,
  new_target = NULL,
  new_type = NULL,
  drop_original_target = FALSE,
  ...
)
}
\arguments{
\item{intask}{\code{\link[mlr3:Task]{Task}}\cr
A \code{\link[mlr3:Task]{Task}} to be converted.}

\item{new_target}{\code{character(1)}\cr
New target to be set, must be a column in the \code{intask} data.
If \code{NULL}, no new target is set, and task is converted as-is.}

\item{new_type}{\code{character(1)}\cr
The new task type. Must be in \code{mlr_reflections$task_types}.
If \code{NULL} (default), a new task with the same task_type is created.}

\item{drop_original_target}{\code{logical(1)}\cr
If \code{FALSE} (default), the original target is added as a feature.
Otherwise the original target is assigned no col_role, i.e. dropped.}

\item{\dots\cr}{Further arguments passed to the constructor of the task.}
}
\value{
\code{\link[mlr3:Task]{Task}}
}
\description{
The task's target is replaced by a different column from the data.
#FIXME: This needs detailed tests.
}
