% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%>>\%}
\alias{\%>>\%}
\title{PipeOp Composition Operator}
\usage{
g1 \%>>\% g2
}
\arguments{
\item{g1}{(\code{\link{Graph}} | \code{\link{PipeOp}}) \cr
\link{Graph} / \link{PipeOp} to put in front of \code{g2}.}

\item{g2}{(\code{\link{Graph}} | \code{\link{PipeOp}}) \cr
\link{Graph} / \link{PipeOp} to put after  \code{g1}.}
}
\description{
This operator \dQuote{pipes} data from the source \code{g1} into the sink \code{g2}. Both source and sink can either be
a \link{Graph} or a \link{PipeOp}. The number of output channels of \code{g1} must equal the number of input channels
of the \code{g2}.
}
