% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\docType{data}
\name{GraphLearner}
\alias{GraphLearner}
\title{GraphLearner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
GraphLearner
}
\description{
A \link{Learner} that encapsulates a \link{Graph} to be used in mlr3
resampling and benchmarks.

The Graph must return a single \code{\link{Prediction}} on its \code{$predict()}
call. The result of the \code{$train()} call is discarded, only the
internal state changes during training are used.
}
\keyword{datasets}
