% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphNode.R
\docType{data}
\name{GraphNode}
\alias{GraphNode}
\title{GraphNode}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GraphNode$new(pipeop)
}
\description{
The \code{GraphNode} class implements behaviors required to define the
connections between the \code{PipeOps}. The set of interconnected
GraphNodes creates the pipeline graph.
}
\keyword{datasets}
