% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greplicate.R
\name{greplicate}
\alias{greplicate}
\title{Create Disjoint Graph Union of Copies of a Graph}
\usage{
greplicate(graph, n)
}
\arguments{
\item{graph}{\code{\link{Graph}} \cr
Graph to replicate.}

\item{n}{\code{integer(1)}
Number of copies to create.}
}
\value{
\code{\link{Graph}} containing \code{n} copies of input \code{graph}.
}
\description{
Create a new \code{\link{Graph}} containing \code{n} copies of the input \code{Graph} / \code{\link{PipeOp}}.
To avoid ID collisions, PipeOp IDs are suffixed with \code{_i}
where \code{i} ranges from 1 to \code{n}.
}
\seealso{
Other Graph operators: \code{\link{\%>>\%}},
  \code{\link{assert_graph}}, \code{\link{assert_pipeop}},
  \code{\link{gunion}}
}
\concept{Graph operators}
