% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunion.R
\name{gunion}
\alias{gunion}
\title{Union of graphs}
\usage{
gunion(graphs)
}
\arguments{
\item{...}{A list of \link{Graph}, \link{GraphNode} or \link{PipeOp} \cr
List of elements with one of the types defined above.}

\item{.graphs}{list of \link{Graph}` \cr
Graphs which are to be joined.}
}
\value{
\code{[Graph]}
}
\description{
Takes an arbitrary amount of Graphs, GraphNodes and PipeOps as inputs and joins
them by "stacking" them one over the other.
Returns the full graph.
}
