% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunion.R
\name{gunion}
\alias{gunion}
\title{Union of graphs}
\usage{
gunion(graphs)
}
\arguments{
\item{graphs}{A list of \link{Graph} or \link{PipeOp} \cr
List of elements with one of the types defined above, which are the
graphs to be joined.}
}
\value{
\code{[Graph]}
}
\description{
Takes an arbitrary amount of \link{Graph}s and/or \link{PipeOp}s as inputs and joins
them by "stacking" them one over the other.
Returns the full graph.
}
