% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunion.R
\name{gunion}
\alias{gunion}
\title{Disjoint Union of Graphs}
\usage{
gunion(graphs)
}
\arguments{
\item{graphs}{\code{list} of (\code{\link{Graph}} | \code{\link{PipeOp}}) \cr
List of elements with one of the types defined above, which are the
graphs to be joined.}
}
\value{
\code{\link{Graph}} the resulting \code{Graph}.
}
\description{
Takes an arbitrary amount of \code{\link{Graph}}s and/or \code{\link{PipeOp}}s as inputs and joins
them in a new \code{Graph}.

The \code{PipeOp}s of the input \code{Graph}s are not joined with new edges across
\code{Graph}s, so if \code{length(graphs) > 1} the resulting \code{Graph} will be disconnected.
}
\seealso{
Other Graph operators: \code{\link{\%>>\%}},
  \code{\link{assert_graph}}, \code{\link{greplicate}}
}
\concept{Graph operators}
