% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3pipelines-package}
\alias{mlr3pipelines}
\alias{mlr3pipelines-package}
\title{mlr3pipelines: Preprocessing Operators and Pipelines for 'mlr3'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Dataflow programming toolkit that enriches 'mlr3' with a diverse set of pipelining operators ('PipeOps') that can be composed into graphs. Operations exist for data preprocessing, model fitting, and ensemble learning. Graphs can themselves be treated as 'mlr3' 'Learners' and can therefore be resampled, benchmarked, and tuned.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3pipelines.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3pipelines}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3pipelines/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
  \item Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Susanne Dandl \email{dandl.susanne@googlemail.com}
}

}
