% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3pipelines-package}
\alias{mlr3pipelines}
\alias{mlr3pipelines-package}
\title{mlr3pipelines: Preprocessing Operators and Pipelines for mlr3}
\description{
Pipelines for mlr3. Work in progress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3pipelines}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3pipelines/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{martin.binder@mail.com}

Other contributors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.net} [contributor]
  \item Michel Lang \email{michellang@gmail.com} [contributor]
  \item Lars Kotthoff \email{larsko@cs.ubc.ca} [contributor]
}

}
\keyword{internal}
