% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3pipelines-package}
\alias{mlr3pipelines}
\alias{mlr3pipelines-package}
\title{mlr3pipelines: Preprocessing Operators and Pipelines for 'mlr3'}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Pipelines for mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3pipelines.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3pipelines}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3pipelines/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (0000-0001-8867-762X)
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)
  \item Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)
  \item Susanne Dandl \email{dandl.susanne@googlemail.com}
}

}
