% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterEnsemble.R
\name{mlr_filters_ensemble}
\alias{mlr_filters_ensemble}
\alias{FilterEnsemble}
\title{Filter Ensemble}
\format{
\code{\link[R6:R6Class]{R6Class}} object inheriting from \code{\link[mlr3filters:Filter]{Filter}}.
}
\description{
Implements the filte rensemble proposed in Binder (2020).
}
