% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{mlr_graphs_bagging}
\alias{mlr_graphs_bagging}
\alias{bagging_pipeline}
\title{Create a bagging learner}
\usage{
bagging_pipeline(graph, iterations = 10, averager = NULL)
}
\arguments{
\item{graph}{\code{\link{PipeOp}} | \code{\link{Graph}} \cr
A learner to create a robustifying pipeline for. Optional, if omitted,
a more conservative pipeline is built.}

\item{iterations}{\code{\link{integer}} \cr
Number of bagging iterations. Defaults to 10.}

\item{averager}{\code{\link{PipeOp}} | \code{\link{Graph}} \cr
A \code{\link{PipeOp}} or \code{\link{Graph}} that averages the predictions from the
replicated and subsampled graph's.
In the simplest case, \code{po("classifavg")} and \code{po("regravg")} can be used
in order to perform simple averaging of classification and regression
predictions respectively.`}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} that performs bagging on the supplied graph.
}
\examples{
library(mlr3)
lrn_po = po("learner", lrn("regr.rpart"))
task = mlr_tasks$get("boston_housing")
gr = bagging_pipeline(lrn_po, 3, averager = po("regravg"))
resample(task, GraphLearner$new(gr), rsmp("holdout"))
}
