% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{mlr_graphs_branching}
\alias{mlr_graphs_branching}
\alias{pipeline_branch}
\title{Create a branch - unbranch pipeline}
\usage{
pipeline_branch(graphs, id_prefix = "", prefix_branches = FALSE)
}
\arguments{
\item{graphs}{List of \code{\link{PipeOp}} | \code{\link{Graph}} \cr
A (possibly named) list of \code{\link{PipeOp}}s or \code{\link{Graph}}s.}

\item{id_prefix}{\code{\link{character(1)}} \cr
Optional id prefix to prepend to \code{\link{PipeOpBranch}} and \code{\link{PipeOpUnbranch}}.
IDs will be \code{"branch"} and \code{"unbranch"}, prefixed by \code{id_prefix}. Default is \code{""}.}

\item{`prefix_branches`}{\code{\link{logical(1)}} \cr
Whether to add \code{id_prefix} as prefix to the \code{branch.selection} when performing \code{gunion}.
Can be helpful to avoid ID clashes in resulting graph. Default \code{FALSE}.}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} that provides branching and unbranching
(using \code{\link{PipeOpBranch}} and \code{\link{PipeOpBranch}})
for a list of supplied \code{\link{PipeOp}}s or \code{\link{Graph}}s.
Takes either the names of the list elements, or the graph's ids
as 'option' for \code{\link{PipeOpBranch}}.
}
\examples{
library(mlr3)
library(mlr3misc)
lrns = map(list(lrn("classif.rpart"), lrn("classif.featureless")), po)
task = mlr_tasks$get("boston_housing")
gr = pipeline_branch(lrns)
# change parameters
# gr$param_set$values$branch.selection = "classif.featureless"
}
