% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_graphs_elements.R
\name{mlr_graphs_targettrafo}
\alias{mlr_graphs_targettrafo}
\alias{pipeline_targettrafo}
\title{Transform and Re-Transform the Target Variable}
\usage{
pipeline_targettrafo(
  graph,
  trafo_pipeop = PipeOpTargetMutate$new(),
  id_prefix = ""
)
}
\arguments{
\item{graph}{\code{\link{PipeOpLearner}} | \code{\link{Graph}} \cr
A \code{\link{PipeOpLearner}} or \code{\link{Graph}} to create a robustifying pipeline for. If this is a \code{\link{Graph}},
the last \code{\link{PipeOp}} should be a \code{\link{PipeOpLearner}} and the first \code{\link{PipeOp}} should accept a
single \code{\link[mlr3:Task]{Task}} as input.}

\item{trafo_pipeop}{\code{\link{PipeOp}} \cr
A \code{\link{PipeOp}} that is a subclass of \code{\link{PipeOpTargetTrafo}}. Default is
\code{\link{PipeOpTargetMutate}}.}

\item{id_prefix}{\code{character(1)} \cr
Optional id prefix to prepend to \code{\link{PipeOpTargetInvert}} ID. The resulting ID will be
\code{"[id_prefix]targetinverter"}. Default is \code{""}.}
}
\value{
\code{\link{Graph}}
}
\description{
Wraps a \code{\link{Graph}} that transforms a target during training and inverts the transformation
during prediction. This is done as follows:
\itemize{
\item Specify a transformation and inversion function using any subclass of \code{\link{PipeOpTargetTrafo}}, defaults to
\code{\link{PipeOpTargetMutate}}, afterwards apply \code{graph}.
\item At the very end, during prediction the transformation is inverted using \code{\link{PipeOpTargetInvert}}.
\item To set a transformation and inversion function for \code{\link{PipeOpTargetMutate}} see the
parameters \code{trafo} and \code{inverter} of the \code{param_set} of the resulting \code{\link{Graph}}.
}
}
\examples{
library("mlr3")

tt = pipeline_targettrafo(PipeOpLearner$new(LearnerRegrRpart$new()))
tt$param_set$values$targetmutate.trafo = function(x) log(x, base = 2)
tt$param_set$values$targetmutate.inverter = function(x) 2 ^ x

# gives the same as
g = Graph$new()
g$add_pipeop(PipeOpTargetMutate$new(param_vals = list(
  trafo = function(x) log(x, base = 2),
  inverter = function(x) 2 ^ x)
  )
)
g$add_pipeop(LearnerRegrRpart$new())
g$add_pipeop(PipeOpTargetInvert$new())
g$add_edge(src_id = "targetmutate", dst_id = "targetinvert",
  src_channel = 1, dst_channel = 1)
g$add_edge(src_id = "targetmutate", dst_id = "regr.rpart",
  src_channel = 2, dst_channel = 1)
g$add_edge(src_id = "regr.rpart", dst_id = "targetinvert",
  src_channel = 1, dst_channel = 2)
}
