% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\docType{data}
\name{mlr_learners_graph}
\alias{mlr_learners_graph}
\alias{GraphLearner}
\title{GraphLearner}
\format{\code{\link{R6Class}} object inheriting from \code{\link[mlr3:Learner]{mlr3::Learner}}.}
\usage{
GraphLearner
}
\description{
A \code{\link[mlr3:Learner]{Learner}} that encapsulates a \code{\link{Graph}} to be used in
\link[mlr3:mlr3-package]{mlr3} resampling and benchmarks.

The Graph must return a single \code{\link[mlr3:Prediction]{Prediction}} on its \code{$predict()}
call. The result of the \code{$train()} call is discarded, only the
internal state changes during training are used.

Note the \code{predict_type} of a \code{\link{GraphLearner}} does currently not track the \code{predict_type}
of any \code{\link[mlr3:Learner]{Learner}} encapsulated within the \code{\link{Graph}}. Therefore, when requesting
e.g. \code{"prob"} predictions, the \code{predict_type} of \emph{both} the encapsulated \code{\link[mlr3:Learner]{Learner}}
and the wrapping \code{\link{GraphLearner}} need to be set to \code{"prob"}.
}
\seealso{
Other Learners: \code{\link{mlr_learners_avg}}
}
\concept{Learners}
\keyword{datasets}
