% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\docType{data}
\name{mlr_learners_graph}
\alias{mlr_learners_graph}
\alias{GraphLearner}
\title{GraphLearner}
\format{\code{\link{R6Class}} object inheriting from \code{\link[mlr3:Learner]{mlr3::Learner}}.}
\usage{
GraphLearner
}
\description{
A \code{\link[mlr3:Learner]{Learner}} that encapsulates a \code{\link{Graph}} to be used in
\link[mlr3:mlr3-package]{mlr3} resampling and benchmarks.

The Graph must return a single \code{\link[mlr3:Prediction]{Prediction}} on its \code{$predict()}
call. The result of the \code{$train()} call is discarded, only the
internal state changes during training are used.
}
\seealso{
Other Learners: \code{\link{mlr_learners_classif.avg}}
}
\concept{Learners}
\keyword{datasets}
