% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEncode.R
\docType{data}
\name{mlr_pipeop_encode}
\alias{mlr_pipeop_encode}
\alias{PipeOpEncode}
\title{PipeOpEncode}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreprocSimple}}.}
\usage{
PipeOpEncode
}
\description{
Encodes \code{factor}, \code{character} and \code{ordered factor} columns.

Possible encodings are \code{"one-hot"} encoding, as well as encoding according to \code{stats::contr.helmert()}, \code{stats::contr.poly()},
\code{stats::contr.sum()} and \code{stats::contr.treatment()}. Newly created columns are named by factor levels in \code{"one-hot"} and
\code{"treatment"} encoding, and numbered otherwise.

Use the \code{$affect_columns} functionality to only encode a subset of columns, or only encode columns of a certain type.
}
\section{Parameter Set}{

\itemize{
\item \code{method}  :: \code{character(1)} \cr
If set to \code{"one-hot"}, creates a new column for each factor level. If set to \code{"treatment"}, creates $n-1$ columns leaving
out the first factor level of each factor variable (see \code{stats::contr.treatment()}). If set to \code{"helmert"}, creates
columns according to Helmert contrasts (see \code{stats::contr.helmert()}). If set to \code{"poly"}, creates columns with
contrasts based on orthogonal polynomials (see \code{stats::contr.poly()}). If set to \code{"sum"}, creates columns with contrasts
summing to zero, (see \code{stats::contr.sum()}).
`
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
