% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{mlr_pipeop_majorityvote}
\alias{mlr_pipeop_majorityvote}
\alias{PipeOpMajorityVote}
\title{PipeOpMajorityVote}
\format{\link{R6Class} PipeOpMajorityVote}
\usage{
PipeOpMajorityVote
}
\description{
Computes the mode over different predictions for each row_id.
Returns a single \code{\link{Prediction}}.
Used for classification \code{Prediction}s.
}
\examples{
op = PipeOpMajorityVote$new(3)
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeop_undersample}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
