% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNULL.R
\docType{data}
\name{mlr_pipeop_NULL}
\alias{mlr_pipeop_NULL}
\alias{PipeOpNULL}
\title{PipeOpNULL}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpNULL
}
\description{
Simply pushes the input forward unchanged.
Can be usefull to keep the original task in conjunction with
\code{gunion()}.
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeop_undersample}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
