% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpUndersample.R
\docType{data}
\name{mlr_pipeop_undersample}
\alias{mlr_pipeop_undersample}
\alias{PipeOpUndersample}
\title{PipeOpUndersample}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}.}
\usage{
PipeOpUndersample
}
\description{
Undersamples (filters) a \code{\link{Task}} to keep only a fraction of the rows with majority class.

Possible parameters are \code{$frac} and \code{$ratio}, only one of which must be given. The majority
class is never "up"-sampled, so if \code{$ratio} times the mean minority class size turns out
larger than the initial majority class size, the number of rows is shuffled but not changed.
}
\section{Parameter Set}{

\itemize{
\item \code{frac}  :: \code{numeric(1)} \cr
Fraction of rows of the majority class to keep in the task.
\item \code{ratio} :: \code{numeric(1)} \cr
Ratio of number of rows of majority class to keep, relative
to average minority class row number. Default is 1.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
