% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_pipeops.R
\name{mlr_pipeops}
\alias{mlr_pipeops}
\title{Dictionary of PipeOps}
\format{\link{R6Class} object}
\description{
A simple \link{Dictionary} storing objects of class \link{PipeOp}.
Each learner has an associated help page, see \code{mlr_pipeops_[id]}.
}
\section{Usage}{


See \link[mlr3:Dictionary]{mlr3::Dictionary}.
}

\examples{
mlr_pipeops$ids()
as.data.table(mlr_pipeops)
}
\concept{Dictionary}
\concept{PipeOps}
