% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_pipeops.R
\docType{data}
\name{mlr_pipeops}
\alias{mlr_pipeops}
\title{Dictionary of PipeOps}
\format{\code{\link{R6Class}} object inheriting from \code{\link[mlr3misc:Dictionary]{mlr3misc::Dictionary}}}
\usage{
mlr_pipeops
}
\description{
A simple \code{\link[mlr3misc:Dictionary]{Dictionary}} storing objects of class \code{\link{PipeOp}}.
Each \code{PipeOp} has an associated help page, see \code{mlr_pipeop_[id]}.
}
\section{Usage}{


See \code{\link[mlr3misc:Dictionary]{mlr3misc::Dictionary}}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{data.table()} with fields "key", "packages", and connectors.
}
}

\seealso{
Other mlr3pipelines backend related: \code{\link{Graph}},
  \code{\link{PipeOpTaskPreprocSimple}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}}

Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_apply}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_mutate}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_select}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}}
}
\concept{PipeOps}
\concept{mlr3pipelines backend related}
\keyword{datasets}
