% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpKernelPCA.R
\docType{data}
\name{mlr_pipeops_kernelpca}
\alias{mlr_pipeops_kernelpca}
\alias{PipeOpKernelPCA}
\title{PipeOpKernelPCA}
\format{\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}/\code{\link{PipeOp}}.}
\description{
Extracts kernel principle components from data. Only affects numerical features.
See \link[kernlab:kpca]{kernlab::kpca} for details.
}
\section{Construction}{
\preformatted{PipeOpKernelPCA$new(id = "kernelpca", param_vals = list())
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier of resulting object, default \code{"kernelpca"}.
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default \code{list()}.
}
}

\section{Input and Output Channels}{

Input and output channels are inherited from \code{\link{PipeOpTaskPreproc}}.

The output is the input \code{\link[mlr3:Task]{Task}} with all affected numeric parameters replaced by their principal components.
}

\section{State}{

The \code{$state} is a named \code{list} with the \code{$state} elements inherited from \code{\link{PipeOpTaskPreproc}},
as well as the returned \code{\link{S4}} object of the function \code{\link[kernlab:kpca]{kernlab::kpca()}}.

The \code{@rotated} slot of the \code{"kpca"} object is overwritten with an empty matrix for memory efficiency.

The slots of the \code{\link{S4}} object can be accessed by accessor function. See \link[kernlab:kpca]{kernlab::kpca}.
}

\section{Parameters}{

The parameters are the parameters inherited from \code{\link{PipeOpTaskPreproc}}, as well as:
\itemize{
\item \code{kernel} :: \code{character(1)}\cr
The standard deviations of the principal components. See \code{\link[kernlab:kpca]{kpca()}}.
\item \code{kpar} :: \code{list}\cr
List of hyper-parameters that are used with the kernel function. See \code{\link[kernlab:kpca]{kpca()}}.
\item \code{features} :: \code{numeric(1)}\cr
Number of principal components to return. Default 0 means that all
principal components are returned. See \code{\link[kernlab:kpca]{kpca()}}.
\item \code{th} :: \code{numeric(1)}\cr
The value of eigenvalue under which principal components are ignored. Default is 0.0001. See \code{\link[kernlab:kpca]{kpca()}}.
\item \code{na.action} :: \code{function}\cr
Function to specify NA action. Default is \code{\link{na.omit}}. See \code{\link[kernlab:kpca]{kpca()}}.
}
}

\section{Internals}{

Uses the \code{\link[kernlab:kpca]{kpca()}} function.
}

\section{Methods}{

Only methods inherited from \code{\link{PipeOpTaskPreproc}}/\code{\link{PipeOp}}.
}

\examples{
library("mlr3")

task = tsk("iris")
pop = po("kernelpca", features = 3)  # only keep top 3 components

task$data()
pop$train(list(task))[[1]]$data()
}
\seealso{
Other PipeOps: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpImpute}},
  \code{\link{PipeOpTaskPreproc}}, \code{\link{PipeOp}},
  \code{\link{mlr_pipeops_boxcox}},
  \code{\link{mlr_pipeops_branch}},
  \code{\link{mlr_pipeops_chunk}},
  \code{\link{mlr_pipeops_classbalancing}},
  \code{\link{mlr_pipeops_classifavg}},
  \code{\link{mlr_pipeops_classweights}},
  \code{\link{mlr_pipeops_colapply}},
  \code{\link{mlr_pipeops_collapsefactors}},
  \code{\link{mlr_pipeops_copy}},
  \code{\link{mlr_pipeops_datefeatures}},
  \code{\link{mlr_pipeops_encodeimpact}},
  \code{\link{mlr_pipeops_encodelmer}},
  \code{\link{mlr_pipeops_encode}},
  \code{\link{mlr_pipeops_featureunion}},
  \code{\link{mlr_pipeops_filter}},
  \code{\link{mlr_pipeops_fixfactors}},
  \code{\link{mlr_pipeops_histbin}},
  \code{\link{mlr_pipeops_ica}},
  \code{\link{mlr_pipeops_imputehist}},
  \code{\link{mlr_pipeops_imputemean}},
  \code{\link{mlr_pipeops_imputemedian}},
  \code{\link{mlr_pipeops_imputenewlvl}},
  \code{\link{mlr_pipeops_imputesample}},
  \code{\link{mlr_pipeops_learner}},
  \code{\link{mlr_pipeops_missind}},
  \code{\link{mlr_pipeops_modelmatrix}},
  \code{\link{mlr_pipeops_mutate}},
  \code{\link{mlr_pipeops_nop}},
  \code{\link{mlr_pipeops_pca}},
  \code{\link{mlr_pipeops_quantilebin}},
  \code{\link{mlr_pipeops_regravg}},
  \code{\link{mlr_pipeops_removeconstants}},
  \code{\link{mlr_pipeops_scalemaxabs}},
  \code{\link{mlr_pipeops_scalerange}},
  \code{\link{mlr_pipeops_scale}},
  \code{\link{mlr_pipeops_select}},
  \code{\link{mlr_pipeops_smote}},
  \code{\link{mlr_pipeops_spatialsign}},
  \code{\link{mlr_pipeops_subsample}},
  \code{\link{mlr_pipeops_unbranch}},
  \code{\link{mlr_pipeops_yeojohnson}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
