% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpOVR.R
\name{mlr_pipeops_ovrunite}
\alias{mlr_pipeops_ovrunite}
\alias{PipeOpOVRUnite}
\title{PipeOpOVRUnite}
\format{
\code{\link{R6Class}} inheriting from \code{\link{PipeOpEnsemble}}/\code{\link{PipeOp}}.
FIXME: document
}
\description{
PipeOpOVRUnite
}
