% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpUMAP.R
\name{mlr_pipeops_umap}
\alias{mlr_pipeops_umap}
\alias{PipeOpUMAP}
\title{Uniform Manifold Approximation and Projection (UMAP)}
\format{
\code{\link{R6Class}} object inheriting from \code{\link{PipeOpTaskPreproc}}/\code{\link{PipeOp}}.
}
\description{
Carry out dimensionality reduction of a dataset using the Uniform Manifold Approximation and Projection (UMAP).
See \code{\link[uwot:umap2]{uwot::umap2()}} for details.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{PipeOpUMAP$new(id = "umap", param_vals = list())
}\if{html}{\out{</div>}}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier of resulting object, default \code{"umap"}.
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default \code{list()}.
}
}

\section{Input and Output Channels}{

Input and output channels are inherited from \code{\link{PipeOpTaskPreproc}}.

The output is the input \code{\link[mlr3:Task]{Task}} with all affected numeric features replaced by their principal components.
}

\section{State}{

The \verb{$state} is a named \code{list} with the \verb{$state} elements inherited from \code{\link{PipeOpTaskPreproc}}, as well as the elements of the class \link[stats:prcomp]{stats::prcomp},
with the exception of the \verb{$x} slot. These are in particular:
\itemize{
\item \code{sdev} :: \code{numeric}\cr
The standard deviations of the principal components.
\item \code{rotation} :: \code{matrix}\cr
The matrix of variable loadings.
\item \code{center} :: \code{numeric} | \code{logical(1)}\cr
The centering used, or \code{FALSE}.
\item \code{scale} :: \code{numeric} | \code{logical(1)}\cr
The scaling used, or \code{FALSE}.
}
}

\section{Parameters}{

The parameters are the parameters inherited from \code{\link{PipeOpTaskPreproc}}, as well as:
\itemize{
\item \code{n_neighbors} :: \code{integer(1)}\cr
Blah
\item \code{n_components} :: \code{integer(1)}\cr
Blah
\item \code{metric} :: \code{character(1)}\cr
Blah
\item \code{n_epochs} :: \code{integer(1)}\cr
Blah
\item \code{learning_rate} :: \code{numeric(1)}\cr
Blah
\item \code{init} :: \code{character(1)}\cr
Blah
\item \code{init_sdev} :: \code{character(1)}\cr
Blah
\item \code{spread} :: \code{character(1)}\cr
Blah
\item \code{min_dist} :: \code{character(1)}\cr
Blah
\item \code{set_op_mix_ratio} :: \code{character(1)}\cr
Blah
\item \code{local_connectivity} :: \code{character(1)}\cr
Blah
\item \code{bandwidth} :: \code{character(1)}\cr
Blah
\item \code{repulsion_strength} :: \code{character(1)}\cr
Blah
\item \code{a} :: \code{character(1)}\cr
Blah
\item \code{b} :: \code{character(1)}\cr
Blah
\item \code{nn_method} :: \code{character(1)}\cr
Blah
\item \code{n_trees} :: \code{character(1)}\cr
Blah
\item \code{search_k} :: \code{character(1)}\cr
Blah
\item \code{approx_pow} :: \code{character(1)}\cr
Blah
\item \code{y} :: \code{character(1)}\cr
Blah
\item \code{target_n_neighbors} :: \code{character(1)}\cr
Blah
\item \code{target_metric} :: \code{character(1)}\cr
Blah
\item \code{target_weight} :: \code{character(1)}\cr
Blah
\item \code{pca} :: \code{character(1)}\cr
Blah
\item \code{pca_center} :: \code{character(1)}\cr
Blah
\item \code{pca_rand} :: \code{character(1)}\cr
Blah
\item \code{fast_sgd} :: \code{character(1)}\cr
Blah
\item \code{n_threads} :: \code{character(1)}\cr
Blah
\item \code{n_sgd_threads} :: \code{character(1)}\cr
Blah
\item \code{grain_size} :: \code{character(1)}\cr
Blah
\item \code{verbose} :: \code{character(1)}\cr
Blah
\item \code{batch} :: \code{character(1)}\cr
Blah
\item \code{opt_args} :: \code{character(1)}\cr
Blah
\item \code{epoch_callback} :: \code{character(1)}\cr
Blah
\item \code{pca_method} :: \code{character(1)}\cr
Blah
\item \code{binary_edge_weights} :: \code{character(1)}\cr
Blah
\item \code{dens_scale} :: \code{character(1)}\cr
Blah
\item \code{seed} :: \code{character(1)}\cr
Blah
\item \code{nn_args} :: \code{character(1)}\cr
Blah
}
}

\section{Internals}{

Uses the \code{\link[uwot:umap]{umap()}} function.
}

\section{Methods}{

Only methods inherited from \code{\link{PipeOpTaskPreproc}}/\code{\link{PipeOp}}.
}

\examples{
library("mlr3")

task = tsk("iris")
pop = po("umap")

task$data()
pop$train(list(task))[[1]]$data()

pop$state
}
\seealso{
https://mlr-org.com/pipeops.html

Other PipeOps: 
\code{\link{PipeOp}},
\code{\link{PipeOpEnsemble}},
\code{\link{PipeOpImpute}},
\code{\link{PipeOpTargetTrafo}},
\code{\link{PipeOpTaskPreproc}},
\code{\link{PipeOpTaskPreprocSimple}},
\code{\link{mlr_pipeops}},
\code{\link{mlr_pipeops_boxcox}},
\code{\link{mlr_pipeops_branch}},
\code{\link{mlr_pipeops_chunk}},
\code{\link{mlr_pipeops_classbalancing}},
\code{\link{mlr_pipeops_classifavg}},
\code{\link{mlr_pipeops_classweights}},
\code{\link{mlr_pipeops_colapply}},
\code{\link{mlr_pipeops_collapsefactors}},
\code{\link{mlr_pipeops_colroles}},
\code{\link{mlr_pipeops_copy}},
\code{\link{mlr_pipeops_datefeatures}},
\code{\link{mlr_pipeops_encode}},
\code{\link{mlr_pipeops_encodeimpact}},
\code{\link{mlr_pipeops_encodelmer}},
\code{\link{mlr_pipeops_featureunion}},
\code{\link{mlr_pipeops_filter}},
\code{\link{mlr_pipeops_fixfactors}},
\code{\link{mlr_pipeops_histbin}},
\code{\link{mlr_pipeops_ica}},
\code{\link{mlr_pipeops_imputeconstant}},
\code{\link{mlr_pipeops_imputehist}},
\code{\link{mlr_pipeops_imputelearner}},
\code{\link{mlr_pipeops_imputemean}},
\code{\link{mlr_pipeops_imputemedian}},
\code{\link{mlr_pipeops_imputemode}},
\code{\link{mlr_pipeops_imputeoor}},
\code{\link{mlr_pipeops_imputesample}},
\code{\link{mlr_pipeops_kernelpca}},
\code{\link{mlr_pipeops_learner}},
\code{\link{mlr_pipeops_missind}},
\code{\link{mlr_pipeops_modelmatrix}},
\code{\link{mlr_pipeops_multiplicityexply}},
\code{\link{mlr_pipeops_multiplicityimply}},
\code{\link{mlr_pipeops_mutate}},
\code{\link{mlr_pipeops_nmf}},
\code{\link{mlr_pipeops_nop}},
\code{\link{mlr_pipeops_ovrsplit}},
\code{\link{mlr_pipeops_ovrunite}},
\code{\link{mlr_pipeops_pca}},
\code{\link{mlr_pipeops_proxy}},
\code{\link{mlr_pipeops_quantilebin}},
\code{\link{mlr_pipeops_randomprojection}},
\code{\link{mlr_pipeops_randomresponse}},
\code{\link{mlr_pipeops_regravg}},
\code{\link{mlr_pipeops_removeconstants}},
\code{\link{mlr_pipeops_renamecolumns}},
\code{\link{mlr_pipeops_replicate}},
\code{\link{mlr_pipeops_scale}},
\code{\link{mlr_pipeops_scalemaxabs}},
\code{\link{mlr_pipeops_scalerange}},
\code{\link{mlr_pipeops_select}},
\code{\link{mlr_pipeops_smote}},
\code{\link{mlr_pipeops_spatialsign}},
\code{\link{mlr_pipeops_subsample}},
\code{\link{mlr_pipeops_targetinvert}},
\code{\link{mlr_pipeops_targetmutate}},
\code{\link{mlr_pipeops_targettrafoscalerange}},
\code{\link{mlr_pipeops_textvectorizer}},
\code{\link{mlr_pipeops_threshold}},
\code{\link{mlr_pipeops_tunethreshold}},
\code{\link{mlr_pipeops_unbranch}},
\code{\link{mlr_pipeops_updatetarget}},
\code{\link{mlr_pipeops_vtreat}},
\code{\link{mlr_pipeops_yeojohnson}}
}
\concept{PipeOps}
