% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO_OP.R
\docType{data}
\name{NO_OP}
\alias{NO_OP}
\title{No-Op Sentinel Used for Alternative Branching}
\format{\code{\link{R6}} object.}
\usage{
NO_OP
}
\description{
Special data type for no-ops. Distinct from \code{NULL} for easier debugging
and distinction from unintentional \code{NULL} returns.
}
\seealso{
Other Path Branching: \code{\link{filter_noop}},
  \code{\link{is_noop}}, \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_unbranch}}
}
\concept{Path Branching}
\keyword{datasets}
