% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{OptimizerGenSA}
\alias{OptimizerGenSA}
\title{Optimizer from GenSA}
\format{\code{\link{R6Class}} object inheriting from \code{\link{Optimizer}}.}
\description{
Optimizer GenSA, internally calls \code{GenSA::GenSA} with parameters specified in
\code{param_vals}.
}
\section{Construction}{
\preformatted{OptimizerGenSA$new(param_vals = list())
}
\itemize{
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings given in \code{param_set}. The
subclass should have its own \code{param_vals} parameter and pass it on to \code{super$initialize()}. Default \code{list()}.
}
}

\section{Methods}{

\itemize{
\item \code{optimize}, see \code{\link{Optimizer}} for details.
}
}

\seealso{
Other Optimizer: 
\code{\link{OptimizerNloptr}},
\code{\link{Optimizer}}
}
\concept{Optimizer}
