% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{OptimizerNloptr}
\alias{OptimizerNloptr}
\title{Optimizer from nloptr}
\format{
\code{\link{R6Class}} object inheriting from \code{\link{Optimizer}}.
}
\description{
Non-linear optimization using \code{nloptr::nloptr}, called with parameters specified in
param_vals`.
}
\section{Construction}{
\preformatted{OptimizerNloptr$new(param_vals = list())
}
\itemize{
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings given in \code{param_set}. The
subclass should have its own \code{param_vals} parameter and pass it on to \code{super$initialize()}. Default \code{list()}.
}
}

\section{Methods}{

\itemize{
\item \code{optimize}, see \code{\link{Optimizer}} for details.
}
}

\seealso{
Other Optimizer: 
\code{\link{OptimizerGenSA}},
\code{\link{Optimizer}}
}
\concept{Optimizer}
