% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpChoice.R
\docType{data}
\name{PipeOpChoice}
\alias{PipeOpChoice}
\title{PipeOpChoice}
\format{\link{R6Class} PipeOpChoice}
\usage{
PipeOpChoice
}
\description{
This is one choice PipeOp! Used for multiplexing between different
possible paths.
}
\examples{
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpChoice$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnchoice$new(choices)

}
\seealso{
Other PipeOp: \code{\link{PipeOpDT}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnchoice}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
