% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpCopy.R
\name{PipeOpCopy}
\alias{PipeOpCopy}
\title{PipeOpCopy}
\format{\link{R6Class} PipeOpCopy}
\description{
Copies its input \code{outnum} times.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = PipeOpCopy$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpCopy}
}
}

\section{Details}{

\itemize{
\item \code{outnum}: \code{integer(1)} Number of times the input is copied.
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpDT}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}

Other PipeOpBroadcast: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}
}
\concept{PipeOp}
\concept{PipeOpBroadcast}
