% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDownsample.R
\name{PipeOpDownsample}
\alias{PipeOpDownsample}
\title{PipeOpDownsample}
\format{\link{R6Class} PipeOpDownsample}
\description{
Subsamples a \link{Task} to use only a fraction of the rows.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = pipeOpDownsample$new(id)} \cr
}
}

\section{Details}{

\itemize{
\item \code{frac}: \code{numeric(1)} Fracentage of rows in the task to keep. Default 1.
\item \code{stratify}: \code{logical(1)} Should the subsamples be stratified by target? Default FALSE.
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpTaskPreprocSimple}},
  \code{\link{PipeOpTaskPreproc}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
