% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDT.R
\name{PipeOpDT}
\alias{PipeOpDT}
\title{PipeOpDT}
\format{\link{R6Class} PipeOpDT}
\description{
This let's us work with a \link{data.table} instead of a Task and
delegate all handling of \link{Task}s and \link{DataBackend} to the class.
Allows us to specify functions  \code{train_dt()} and \code{predict_dt} instead of
\code{train()} and \code{predict}, that expect the \link{data.table} containing only
the features from a \link{Task}, and automatically
reconstruct the appropriate Task from a returned \link{data.table}.
We thus enforce: \link{data.table} -> \link{data.table}.
For examples see \link{PipeOpPCA} or \link{PipeOpScale}.

The underlying operation must not change row order or number.
}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
