% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{PipeOpEnsemble}
\alias{PipeOpEnsemble}
\title{PipeOpEnsemble}
\format{\link{R6Class} PipeOpEnsemble}
\usage{
PipeOpEnsemble
}
\description{
Abstract base-class.
Aggregates a list of predictions.
}
\section{Usage}{

Inherits from \link{PipeOpEnsemble}
\itemize{
\item \code{f = PipeOpEnsemble$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpEnsemble}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\seealso{
Other PipeOp, PipeOpAggregate, PipeOpEnsemble: \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}
}
\concept{PipeOp, PipeOpAggregate, PipeOpEnsemble}
\keyword{datasets}
