% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{PipeOpEnsemble}
\alias{PipeOpEnsemble}
\title{PipeOpEnsemble}
\format{Abstract \code{\link{R6Class}} inheriting from \code{\link{PipeOp}}.}
\usage{
PipeOpEnsemble
}
\description{
Parent class for PipeOps that aggregate a list of predictions.
}
\section{Methods}{

\itemize{
\item \code{PipeOpEnsemble$new(innum, id)} \cr
(\code{numeric(1)}, \code{character(1)}) -> \code{self} \cr
Constructor. \code{innum} determines the number of input channels.
}
}

\seealso{
Other PipeOps: \code{\link{PipeOpTaskPreproc}},
  \code{\link{PipeOp}}, \code{\link{mlr_pipeop_NULL}},
  \code{\link{mlr_pipeop_backup_learner}},
  \code{\link{mlr_pipeop_balancesample}},
  \code{\link{mlr_pipeop_branch}},
  \code{\link{mlr_pipeop_chunk}},
  \code{\link{mlr_pipeop_copy}},
  \code{\link{mlr_pipeop_encode}},
  \code{\link{mlr_pipeop_featureunion}},
  \code{\link{mlr_pipeop_filter}},
  \code{\link{mlr_pipeop_impute}},
  \code{\link{mlr_pipeop_learner}},
  \code{\link{mlr_pipeop_majorityvote}},
  \code{\link{mlr_pipeop_modelavg}},
  \code{\link{mlr_pipeop_pca}},
  \code{\link{mlr_pipeop_scale}},
  \code{\link{mlr_pipeop_subsample}},
  \code{\link{mlr_pipeop_unbranch}},
  \code{\link{mlr_pipeops}}
}
\concept{PipeOps}
\keyword{datasets}
