% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\name{PipeOpEnsemble}
\alias{PipeOpEnsemble}
\title{PipeOpEnsemble}
\format{\link{R6Class} PipeOpEnsemble}
\description{
Abstract base-class.
Aggregates a list of predictions.
}
\section{Usage}{

Inherits from \link{PipeOpEnsemble}
\itemize{
\item \code{f = PipeOpEnsemble$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpEnsemble}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}

Other PipeOpAggregate: \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}},
  \code{\link{PipeOpUnbranch}}

Other PipeOpEnsemble: \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}
}
\concept{PipeOp}
\concept{PipeOpAggregate}
\concept{PipeOpEnsemble}
