% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpLearner.R
\name{PipeOpLearner}
\alias{PipeOpLearner}
\title{PipeOpLearner}
\format{\link{R6Class} PipeOpLearner}
\description{
Wraps a \link[mlr3:Learner]{mlr3::Learner} into a \link{PipeOp}.
Inherits the \code{param_set} from the \link[mlr3:Learner]{mlr3::Learner} it is constructed from.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = PipeOpLearner$new(outnum, id)} \cr
\code{[Learner]} -> \link{PipeOpLearner}
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
