% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\docType{data}
\name{PipeOpMajorityVote}
\alias{PipeOpMajorityVote}
\title{PipeOpMajorityVote}
\format{\link{R6Class} PipeOpMajorityVote}
\usage{
PipeOpMajorityVote
}
\description{
Computes the mode over different predictions for each row_id.
}
\section{Usage}{

Inherits from \link{PipeOpMajorityVote}
\itemize{
\item \code{f = PipeOpMajorityVote$new(outnum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpMajorityVote}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\examples{
op = PipeOpMajorityVote$new(3)
}
\seealso{
Other PipeOp, PipeOpAggregate, PipeOpEnsemble: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpModelAvg}}
}
\concept{PipeOp, PipeOpAggregate, PipeOpEnsemble}
\keyword{datasets}
