% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpEnsemble.R
\name{PipeOpModelAvg}
\alias{PipeOpModelAvg}
\title{PipeOpModelAvg}
\format{\link{R6Class} PipeOpModelAvg}
\description{
Averages its input (a list of \link[mlr3:Prediction]{mlr3::Prediction}).
Returns a \link[mlr3:Prediction]{mlr3::Prediction}.
}
\section{Usage}{

Inherits from \link{PipeOpEnsemble}
\itemize{
\item \code{f = PipeOpModelAvg$new(innum, id)} \cr
\code{integer(1)}, \code{character(1)} -> \link{PipeOpModelAvg}
}
}

\section{Details}{

\itemize{
\item \code{innum}: \code{integer(1)} Number of inputs.
}
}

\examples{
op = PipeOpModelAvg$new(3)
}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnbranch}},
  \code{\link{PipeOp}}

Other PipeOpAggregate: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpUnbranch}}

Other PipeOpEnsemble: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpMajorityVote}}
}
\concept{PipeOp}
\concept{PipeOpAggregate}
\concept{PipeOpEnsemble}
