% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpNULL.R
\name{PipeOpNULL}
\alias{PipeOpNULL}
\title{PipeOpNULL}
\format{\link{R6Class} PipeOpNULL}
\description{
Simply pushes the input forward unchanged.
Can be usefull for example to keep the original task in conjunction with
\code{gunion()} to keep a copy of the original data.
}
\section{Usage}{

Inherits from \link{PipeOp}
\itemize{
\item \code{f = pipeOpNULL$new(id)} \cr
\code{character(1)} -> \link{PipeOpNULL}
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnBranch}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
