% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpPCA.R
\name{PipeOpPCA}
\alias{PipeOpPCA}
\title{PipeOpPCA}
\format{\link{R6Class} PipeOpPCA}
\description{
Extracts principle components from data.
See \link[stats:prcomp]{stats::prcomp} for details  and parameters.
}
\section{Usage}{

Inherits from \link{PipeOpDT}
\itemize{
\item \code{f = pipeOpPCA$new(id)} \cr
\code{character(1)} -> \link{PipeOpPCA}
}
}

\examples{
# Instantiate PipeOpPCA
op1 = PipeOpPCA$new()
}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOpUnBranch}},
  \code{\link{PipeOp}}
}
\concept{PipeOp}
