% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpScale.R
\name{PipeOpScale}
\alias{PipeOpScale}
\title{PipeOpScale}
\format{\link{R6Class} PipeOpScale}
\description{
Scales the data to mean = 0 and standard deviation 1.
See \link[base:scale]{base::scale} for details and parameters.
}
\section{Usage}{

Inherits from \link{PipeOpTaskPreproc}
\itemize{
\item \code{f = pipeOpDT$new(id)} \cr
\code{character(1)} -> \link{PipeOpScale}
}
}

\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}},
  \code{\link{PipeOpTaskPreprocSimple}},
  \code{\link{PipeOpTaskPreproc}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
