% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpTaskPreproc.R
\name{PipeOpTaskPreproc}
\alias{PipeOpTaskPreproc}
\title{PipeOpTaskPreproc}
\format{\link{R6Class} PipeOpTaskPreproc}
\description{
Base class for handling most "preprocessing" operations. These
are operations that have exactly one Task input and one Task output,
and expect the column layout of these tasks during input and output
to be the same.

Users must implement train_task and predict_task, which have a Task
input and should manipulate that task in-place; return values are
ignored.

If \code{can_subset} is \code{TRUE}, then the hyperparameter \code{affect_columns}
is added, which is a function that takes the Task as input and returns
a logical as output, indicating what columns to choose.
}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpMajorityVote}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOpTaskPreprocSimple}},
  \code{\link{PipeOpUnbranch}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
