% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpUnBranch.R
\name{PipeOpUnBranch}
\alias{PipeOpUnBranch}
\title{PipeOpUnbranch}
\format{\link{R6Class} PipeOpUnbranch}
\description{
Used to bring together different paths created by \code{\link{PipeOpBranch}}.
}
\section{Methods}{

\itemize{
\item \code{new(options = 1)} \cr
(\code{integer(1)} | \code{character}) -> \code{\link{PipeOpBranch}}
}
}

\section{Details}{

Creates a PipeOp with multiple input channels that can be used to
create a Graph network with alternative paths. \code{options} works as in \code{\link{PipeOpBranch}}
and should probably be the same value as the \code{options} given to the corresponding
\code{\link{PipeOpBranch}} instance.
}

\examples{
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpUnbranch$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnbranch$new(choices)

}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpChunk}}, \code{\link{PipeOpCopy}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpLearner}},
  \code{\link{PipeOpModelAvg}}, \code{\link{PipeOpNULL}},
  \code{\link{PipeOpPCA}}, \code{\link{PipeOpScale}},
  \code{\link{PipeOp}}

Other PipeOpAggregate: \code{\link{PipeOpEnsemble}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpModelAvg}}
}
\concept{PipeOp}
\concept{PipeOpAggregate}
