% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpBranch.R
\docType{data}
\name{PipeOpUnbranch}
\alias{PipeOpUnbranch}
\title{PipeOpUnbranch}
\format{\link{R6Class} PipeOpUnbranch}
\usage{
PipeOpUnbranch
}
\description{
Used to bring together different paths created by \code{\link{PipeOpBranch}}.
}
\examples{
pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpUnbranch$new(choices) \%>>\% gunion(pca, nop) \%>>\% PipeOpUnbranch$new(choices)

}
\seealso{
Other PipeOp: \code{\link{PipeOpBranch}},
  \code{\link{PipeOpDT}}, \code{\link{PipeOpDownsample}},
  \code{\link{PipeOpFeatureUnion}},
  \code{\link{PipeOpNULL}}, \code{\link{PipeOpPCA}},
  \code{\link{PipeOpScale}}, \code{\link{PipeOp}}
}
\concept{PipeOp}
\keyword{datasets}
