% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_pipeops.R
\name{po}
\alias{po}
\title{Short Function to Construct a PipeOp}
\usage{
po(key, ...)
}
\arguments{
\item{key}{(\code{character(1)})\cr
Key of \code{PipeOp} in \code{mlr_pipeops} dictionary to construct.}

\item{...}{(\code{any})\cr
Further arguments of \code{PipeOp} constructor.}
}
\description{
This is a short form of \code{mlr_pipeops$get()}.
}
