% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po}
\alias{po}
\title{Shorthand for mlr3pipelines::mlr_pipeops$get}
\usage{
po(.key, ...)
}
\arguments{
\item{.key}{\code{[character(1)]}\cr
The pipeop to construct}

\item{...}{\code{any}\cr
Additional parameters to give to constructed object.
This may be an argument of the constructor of the
\code{PipeOp}, in which case it is given to this constructor;
or it may be a parameter value, in which case it is
given to the \code{param_vals} argument of the constructor.}
}
\description{
Create a \code{PipeOp} from \code{mlr_pipeops} from given ID, with
given parameters and \code{param_vals}.
}
\examples{
po("learner", "classif.rpart", cp = 0.3)
# is equivalent with:
mlr3pipelines::mlr_pipeops$get("learner", "classif.rpart",
  param_vals = list(cp = 0.3))
}
