% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{ppl}
\alias{ppl}
\title{Shorthand Graph Constructor}
\usage{
ppl(key, ..., id_prefix = "", id_postfix = "")
}
\arguments{
\item{key}{\verb{[character(1)]}\cr
The key of the \code{\link{Graph}} in \code{\link{mlr_graphs}}.}

\item{...}{\code{any}\cr
Additional parameters to give to constructed object.
This may be an argument of the constructor of the
underlying function.}

\item{id_prefix}{\verb{[character(1)]}\cr
Additional prefixes for \code{\link{PipeOp}}'s existing ids.
Defaults to \code{""}, i.e. no changes.}

\item{id_postfix}{\verb{[character(1)]}\cr
Additional postifxes for \code{\link{PipeOp}}'s existing ids.
Defaults to \code{""}, i.e. no changes.}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} from \code{\link{mlr_graphs}} from given ID
}
\examples{
library("mlr3")

gr = ppl("bagging", graph = po(lrn("regr.rpart")), averager = po("regravg"))
}
