% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{ppl}
\alias{ppl}
\title{Shorthand Graph Constructor}
\usage{
ppl(key, ...)
}
\arguments{
\item{key}{\code{[character(1)]}\cr
The key of the \code{\link{Graph}} in \code{\link{mlr_graphs}}.}

\item{...}{\code{any}\cr
Additional parameters to give to constructed object.
This may be an argument of the constructor of the
underlying function.}
}
\value{
\code{\link{Graph}}
}
\description{
Creates a \code{\link{Graph}} from \code{\link{mlr_graphs}} from given ID
}
\examples{
library("mlr3")

gr = ppl("bagging", graph = po(lrn("regr.rpart")), averager = po("regravg"))
}
