% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Simple Preprocessing}
\usage{
preproc(indata, graph, state = NULL, predict = !is.null(state))
}
\arguments{
\item{indata}{(\code{\link[mlr3:Task]{Task}} | \code{\link[data.table:data.table]{data.table}} )\cr
Data to be pre-processed. If \code{indata} is a \code{data.table}, it is used to
construct a \code{\link[mlr3:TaskUnsupervised]{TaskUnsupervised}} internally, meaning
that it would not work with \code{\link{Graph}}s or \code{\link{PipeOp}}s that use the columns
of role \code{target}.}

\item{graph}{(\code{\link{Graph}} | \code{\link{PipeOp}})\cr
modified-by-reference}

\item{state}{(named \code{list} | \code{NULL})\cr
a}

\item{predict}{(\code{logical(1)})\cr
a}
}
\value{
only single output channel
\itemize{
\item makes sense for piping and nature of this function to simplify things (i.e. should be used in simple context)
\item we couldn't identify the proper output channel (e.g. targetmutate)
}
}
\description{
Simple Preprocessing
}
\examples{
# example code

}
