% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Simple Preprocessing}
\usage{
preproc(indata, graph, state = NULL, predict = !is.null(state))
}
\arguments{
\item{indata}{(\code{Task} | \code{data.frame}-like object)\cr
probably just write data.table. I think this is our convention.

If \code{indata} is a \code{data.frame}-like object, it is used to contruct a \code{TaskUnsupervised} internally.}

\item{graph}{\code{Graph} | \code{PipeOp}\cr
modified-by-reference}

\item{state}{named \code{list} | \code{NULL}\cr
a}

\item{predict}{\code{logical(1)}\cr
a}
}
\value{
description
only single output channel
\itemize{
\item makes sense for piping and nature of this function to simplify things (i.e. should be used in simple context)
\item we couldn't identify the proper output channel (e.g. targetmutate)
}
}
\description{
calling train and predict with single_input = TRUE -> input task is copied and given to each unconnected input channel
}
\examples{
# example code

}
