% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\name{set_validate.GraphLearner}
\alias{set_validate.GraphLearner}
\title{Configure Validation for a GraphLearner}
\usage{
set_validate.GraphLearner(learner, validate, which = NULL)
}
\arguments{
\item{learner}{(\code{\link{GraphLearner}})\cr
The graph learner to configure.}

\item{validate}{(\code{numeric(1)}, \code{"inner_valid"} or \code{NULL})\cr
How to set the \verb{$validate} field of the learner.
If set to \code{NULL} all validation is disabled.}

\item{which}{(\code{NULL} or \code{character()})\cr
For which pipeops to enable validation.
This parameter is ignored when \code{validate} is set to \code{NULL}.
By default, validation is enabled for the base learner.}
}
\description{
Configure validation for a graph learner.

In a \code{\link{GraphLearner}}, validation can be configured on two levels:
\enumerate{
\item On the \code{\link{GraphLearner}} level.
\item On the level of the \code{\link{Learner}}s that are wrapped by \code{\link{PipeOpLearner}} and \code{\link{PipeOpLearnerCV}}.
}

Therefore, enabling validation requires to specify not only how to create the validation set (1), but also for which
pipeops to then actually usse it (2).
Moreover, only the \code{\link{GraphLearner}} can specify \strong{how} to create the validation set.
All \code{\link{PipeOpLearner}} and \code{\link{PipeOpLearnerCV}} can only set it to \code{NULL} (disable) and \code{"inner_valid"} (enable).
}
\examples{
# simple
glrn = as_learner(po("pca") \%>>\% lrn("classif.debug"))
set_validate(glrn, 0.3)
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate
set_validate(glrn, NULL)
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate

# complex
glrn = as_learner(ppl("stacking", lrns(c("classif.debug", "classif.featureless")), lrn("classif.debug", id = "final")))
set_validate(glrn, 0.2, which = c("classif.debug", "final"))
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate
glrn$graph$pipeops$final$learner$validate
}
