% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphLearner.R
\name{set_validate.GraphLearner}
\alias{set_validate.GraphLearner}
\title{Configure Validation for a GraphLearner}
\usage{
\method{set_validate}{GraphLearner}(
  learner,
  validate,
  ids = NULL,
  args_all = list(),
  args = list(),
  ...
)
}
\arguments{
\item{learner}{(\code{\link{GraphLearner}})\cr
The graph learner to configure.}

\item{validate}{(\code{numeric(1)}, \code{"predefined"}, \code{"test"}, or \code{NULL})\cr
How to set the \verb{$validate} field of the learner.
If set to \code{NULL} all validation is disabled, both on the graph learner level, but also for all pipeops.}

\item{ids}{(\code{NULL} or \code{character()})\cr
For which pipeops to enable validation.
This parameter is ignored when \code{validate} is set to \code{NULL}.
By default, validation is enabled for the base learner.}

\item{args_all}{(\code{list()})\cr
Rarely needed. A named list of parameter values that are passed to all subsequet \code{\link[=set_validate]{set_validate()}} calls on the individual
\code{PipeOp}s.}

\item{args}{(named \code{list()})\cr
Rarely needed.
A named list of lists, specifying additional argments to be passed to \code{\link[=set_validate]{set_validate()}} for the respective learners.
Names must be a subset of the \code{ids}.}

\item{...}{(any)\cr
Currently unused.}
}
\description{
Configure validation for a graph learner.

In a \code{\link{GraphLearner}}, validation can be configured on two levels:
\enumerate{
\item On the \code{\link{GraphLearner}} level, which specifies \strong{how} the validation set is constructed before entering the graph.
\item On the level of the \code{\link{Learner}}s that are wrapped by \code{\link{PipeOpLearner}} and \code{\link{PipeOpLearnerCV}}, which specifies
which pipeops actually make use of the validation data.
All learners wrapped by \code{\link{PipeOpLearner}} and \code{\link{PipeOpLearnerCV}} should in almost all cases either set it
to \code{NULL} (disable) or \code{"predefined"} (enable).
}
}
\examples{
library(mlr3)
# simple
glrn = as_learner(po("pca") \%>>\% lrn("classif.debug"))
set_validate(glrn, 0.3)
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate
set_validate(glrn, NULL)
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate

# complex
glrn = as_learner(ppl("stacking", list(lrn("classif.debug"), lrn("classif.featureless")),
  lrn("classif.debug", id = "final")))
set_validate(glrn, 0.2, ids = c("classif.debug", "final"))
glrn$validate
glrn$graph$pipeops$classif.debug$learner$validate
glrn$graph$pipeops$final$learner$validate
}
